### 2021-06-14 #3368 @danielskatz
@derrynknife - thanks for your submission.

Due to a current shortage of editors (that we are working to resolve), I'm going to put it on our waiting list until an appropriate editor is available.

In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
