+++ paper.bib
@@ -8,7 +8,8 @@
   title={Survival analysis: techniques for censored and truncated data},
   author={Klein, John P and Moeschberger, Melvin L},
   year={2006},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi={10.1007/b97377}
 }
 @article{turnbull1976empirical,
   title={The empirical distribution function with arbitrarily grouped, censored and truncated data},
@@ -18,7 +19,8 @@
   number={3},
   pages={290--295},
   year={1976},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.2517-6161.1976.tb01597.x}
 }
 @book{meeker2014statistical,
   title={Statistical methods for reliability data},
@@ -46,7 +48,8 @@
   number={1},
   pages={27--52},
   year={1969},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/00224065.1969.11980344}
 }
 
 @article{aalen1978nonparametric,
@@ -55,7 +58,8 @@
   journal={The Annals of Statistics},
   pages={701--726},
   year={1978},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1214/aos/1176344247}
 }
 
 @book{fleming2011counting,
@@ -99,7 +103,8 @@
   volume={4},
   number={40},
   pages={1317},
-  year={2019}
+  year={2019},
+  doi={10.21105/joss.01317}
 }
 
 @inproceedings{maclaurin2015autograd,
@@ -119,5 +124,6 @@
   number={3},
   pages={293--297},
   year={1951},
-  publisher={New York}
+  publisher={New York},
+  doi={10.1115/1.4010337}
 }
