### 2021-06-17 #3368 @derrynknife
Thanks @danielskatz, I've updated the references and pdf as Whedon suggested.

Totally understand that editing will be harder given the state of affairs in the world!

@dfm looking forward to your comments as well.

Thanks team!
