+++ docs/surpyval.rst
@@ -11,6 +11,7 @@ Subpackages
 
 .. toctree::
 
+   surpyval
    surpyval.datasets
-   surpyval.nonparametric
-   surpyval.parametric
+   .. surpyval.nonparametric
+   .. surpyval.parametric
