### 2021-06-17 1d60583
Added docs badge
