+++ README.md
@@ -21,21 +21,21 @@ This is another pure python survival analysis tool so why was it needed? The int
 | **Fleming-Harrington** | Non-Parametric | Yes | Right only | Left only |
 | **Turnbull** | Non-Parametric | Yes | Yes | Yes |
 
-SurPyval also offers many different distributions for users, and because of the flexible implementation adding new distributions is easy. The currently available distributions are:
+SurPyval also offers many different distributions for users, and because of the flexible implementation adding new distributions is easy. Further, the power of SurPyval lay in the robust parameter estimation, as such, some distributions, those that are supported on the half real line, can be offset to make a three- or four-parameter version. The currently available distributions are:
 
-| Distribution |
-| ---- |
-| **Weibull** |
-| **Normal** |
-| **LogNormal** |
-| **Gamma** |
-| **Beta** |
-| **Uniform** |
-| **Exponential** |
-| **Exponentiated Weibull** |
-| **Gumbel** |
-| **Logistic** |
-| **LogLogistic** |
+| Distribution | Offsetable |
+| ---- | ---- | No |
+| **Weibull** | Yes |
+| **Normal** | No |
+| **LogNormal** | Yes |
+| **Gamma** | Yes |
+| **Beta** | No |
+| **Uniform** | No |
+| **Exponential** | Yes |
+| **Exponentiated Weibull** | Yes |
+| **Gumbel** | No |
+| **Logistic** | No |
+| **LogLogistic** | Yes |
 
 This project spawned from a Reliaility Engineering project; due to the history of reliability engineers estimating parameters from a probability plot. SurPyval has continued this tradition to ensure that any parametric distribution can have the estimate plotted on a probability plot. These visualisations enable an analyst to get a sense of the goodness of fit of the parametric distribution with the non-parametric distribution.
 
