+++ surpyval/parametric/parametric.py
@@ -10,7 +10,7 @@ from surpyval import nonparametric as nonp
 import matplotlib.pyplot as plt
 from matplotlib.ticker import FixedLocator
 
-def round_vals(x):
+def _round_vals(x):
 	not_different = True
 	i = 1
 	while not_different:
@@ -20,40 +20,318 @@ def round_vals(x):
 	return x_ticks
 
 class Parametric():
+	"""
+	Result of ``.fit()`` or ``.from_params()`` method for every surpyval distribution.
+
+	Instances of this class are very useful when a user needs the other functions of a distribution for plotting, optimizations, monte carlo analysis and numeric integration.
+
+	"""
 	def __str__(self):
 		return 'Parametric Surpyval model with {dist} distribution fitted by {method} yielding parameters {params}'.format(dist=self.dist.name, method=self.method, params=self.params)
 
 	def sf(self, x):
+		r"""
+
+		Surival (or Reliability) function for a distribution using the parameters found in the ``.params`` attribute.
+
+		Parameters
+		----------
+
+		x : array like or scalar
+			The values of the random variables at which the survival function will be calculated 
+
+		Returns
+		-------
+
+		sf : scalar or numpy array 
+			The scalar value of the survival function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the survival function at each corresponding value in the input array.
+
+
+		Examples
+		--------
+		>>> from surpyval import Weibull
+		>>> model = Weibull.from_params([10, 3])
+		>>> model.sf(2)
+		0.9920319148370607
+		>>> model.sf([1, 2, 3, 4, 5])
+		array([0.9990005 , 0.99203191, 0.97336124, 0.938005  , 0.8824969 ])
+		"""
+		if type(x) == list:
+			x = np.array(x)
 		return self.dist.sf(x, *self.params)
 
 	def ff(self, x):
+		r"""
+
+		The cumulative distribution function, or failure function, for a distribution using the parameters found in the ``.params`` attribute.
+
+		Parameters
+		----------
+
+		x : array like or scalar
+			The values of the random variables at which the failure function (CDF) will be calculated
+
+		Returns
+		-------
+
+		ff : scalar or numpy array 
+			The scalar value of the CDF of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the CDF at each corresponding value in the input array.
+
+
+		Examples
+		--------
+
+		>>> from surpyval import Weibull
+		>>> model = Weibull.from_params([10, 3])
+		>>> model.ff(2)
+		0.007968085162939342
+		>>> model.ff([1, 2, 3, 4, 5])
+		array([0.0009995 , 0.00796809, 0.02663876, 0.061995  , 0.1175031 ])
+		"""
+		if type(x) == list:
+			x = np.array(x)
 		return self.dist.ff(x, *self.params)
 
 	def df(self, x): 
+		r"""
+
+		The density function for a distribution using the parameters found in the ``.params`` attribute.
+
+		Parameters
+		----------
+
+		x : array like or scalar
+			The values of the random variables at which the density function will be calculated
+
+		Returns
+		-------
+
+		df : scalar or numpy array 
[…]
