+++ surpyval/nonparametric/fleming_harrington.py
@@ -6,7 +6,14 @@ from surpyval.nonparametric.nonparametric_fitter import NonParametricFitter
 def fleming_harrington(x, c=None, n=None, **kwargs):
 	r"""
 
-	Fleming-Harrington estimation of Reliability function.
+	Fleming-Harrington estimation of survival distribution.
+
+	x : array like
+        The values of the random variables.
+	c : array like, default is None
+        The array, of same length as x, that flags whether a value in x is left censored (-1), right censored (1), observed (0), or interval censored (2)
+    n : int
+        The number of trials (copied from `binomtest` input).
 
 	Ref:
 	Fleming, T. R., and Harrington, D. P. (1984). “Nonparametric Estimation of the Survival Distribution in Censored Data.” Communications in Statistics—Theory and Methods 13:2469–2486.
@@ -23,10 +30,17 @@ def fleming_harrington(x, c=None, n=None, **kwargs):
 	.. math:: H = cumsum(h)
 
 	Reliability Function:
+	---------------------
 
 	.. math:: R = e^{-H}
 
-	"""
+
+    Methods
+    -------
+    proportion_ci :
+        Compute the confidence interval for the estimate of the proportion.
+
+    """
 	x, r, d = surpyval.xcnt_to_xrd(x, c, n, **kwargs)
 
 	h = [np.sum([1./(r[i]-j) for j in range(d[i])]) for i in range(len(x))]
