+++ surpyval/parametric/logistic.py
@@ -76,11 +76,11 @@ class Logistic_(ParametricFitter):
 
 	def unpack_rr(self, params, rr):
 		if   rr == 'y':
-			sigma = 1/params[0]
+			sigma = 1./params[0]
 			mu    = -sigma * params[1]
 		elif rr == 'x':
-			sigma  = 1./params[0]
-			mu = np.exp(params[1] / (beta * params[0]))
+			sigma  = params[0]
+			mu = params[1]
 		return mu, sigma
 
 Logistic = Logistic_('Logistic')
