+++ docs/exponential.rst
@@ -1,0 +1,7 @@
+Exponential Distribution
+========================
+
+.. automodule:: surpyval.parametric.exponential
+   :members:
+   :undoc-members:
+   :inherited-members: fit
