+++ docs/SurPyval Models.rst
@@ -556,6 +556,3 @@ This shows the power of the flexible API that surpyval offers, because if your m
 
 
 
-
-
-
