### 2021-07-08 9fb30ba
Fixed bugs identified by flake8 tests
