+++ surpyval/parametric/normal.py
@@ -23,7 +23,7 @@ class Normal_(ParametricFitter):
 			'sigma' : 1
 		}
 
-	def parameter_initialiser(self, x, c=None, n=None, t=None):
+	def _parameter_initialiser(self, x, c=None, n=None, t=None):
 		return para.Normal.fit(x, c, n, t, how='MPP').params
 		# x, c, n = surpyval.xcn_handler(x, c, n)
 		# flag = (c == 0).astype(int)
