+++ surpyval/parametric/lognormal.py
@@ -21,7 +21,7 @@ class LogNormal_(ParametricFitter):
 			'sigma' : 1
 		}
 
-	def parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
+	def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
 		# Need an offset mpp function here
 		norm_mod = para.Normal.fit(np.log(x), c=c, n=n, t=t, how='MLE')
 		mu, sigma = norm_mod.params
