### 2021-07-08 5cc8662
Fixed bug identified by flake8 tests with LFP
