### 2021-07-08 437b410
Fixed bug identified by flake8 tests with Beta
