### 2021-07-08 d235178
Converted test.py to flake8 standard
