+++ surpyval/tests/test_fit.py
@@ -1,147 +1,152 @@
 import pytest
 import numpy as np
-from surpyval import *
+from surpyval import Gumbel, Normal, Weibull, LogNormal
+from surpyval import Logistic, LogLogistic, Uniform
 
 DISTS = [Gumbel, Normal, Weibull, LogNormal, Logistic, LogLogistic, Uniform]
 parameter_sample_bounds = [((1, 20), (0.5, 5)),
-						   ((1, 100), (0.5, 100)),
-						   ((1, 100), (0.5, 20)),
-						   ((1, 3), (0.2, 1)),
-						   ((1, 100), (0.5, 20)),
-						   ((1, 100), (0.5, 20)),
-						   ((1, 100), (1, 100)),
-						   ]
+                           ((1, 100), (0.5, 100)),
+                           ((1, 100), (0.5, 20)),
+                           ((1, 3), (0.2, 1)),
+                           ((1, 100), (0.5, 20)),
+                           ((1, 100), (0.5, 20)),
+                           ((1, 100), (1, 100)),
+                           ]
 FIT_SIZES = [5000, 10000, 20000, 50000]
 
+
 def generate_mle_test_cases():
-	for idx, dist in enumerate(DISTS):
-		bounds = parameter_sample_bounds[idx]
-		for kind in ['full', 'censored', 'truncated', 'interval']:
-			yield dist, bounds, kind
+    for idx, dist in enumerate(DISTS):
+        bounds = parameter_sample_bounds[idx]
+        for kind in ['full', 'censored', 'truncated', 'interval']:
+            yield dist, bounds, kind
+
 
 def generate_mpp_test_cases():
-	for idx, dist in enumerate(DISTS):
-		bounds = parameter_sample_bounds[idx]
-		for rr in ['x', 'y']:
-			yield dist, bounds, rr
+    for idx, dist in enumerate(DISTS):
+        bounds = parameter_sample_bounds[idx]
+        for rr in ['x', 'y']:
+            yield dist, bounds, rr
+
 
 def idfunc(x):
-	if type(x) is tuple:
-		return 'bounds'
-	elif type(x) is str:
-		return x
-	else:
-		return x.name
+    if type(x) is tuple:
+        return 'bounds'
+    elif type(x) is str:
+        return x
+    else:
+        return x.name
 
 
 def interval_censor(x, n=100):
-	n, xx = np.histogram(x, bins=n)
-	x = np.vstack([xx[0:-1], xx[1::]]).T
-	x = x[n > 0]
-	n = n[n > 0]
-	return x, n
+    n, xx = np.histogram(x, bins=n)
+    x = np.vstack([xx[0:-1], xx[1::]]).T
+    x = x[n > 0]
+    n = n[n > 0]
+    return x, n
+
 
 def censor_at(x, q, where='right'):
-	c = np.zeros_like(x)
-	x = np.copy(x)
-	if where == 'right':
-		x_q = np.quantile(x, 1 - q)
-		mask = x > x_q
-		c[mask] = 1
-		x[mask] = x_q
-		return x, c
-	elif where == 'left':
-		x_q = np.quantile(x, q)
-		mask = x < x_q
-		c[mask] = -1
-		x[mask] = x_q
-		return x, c
-	elif where == 'both':
-		x_u = np.quantile(x, 1 - q)
-		x_l = np.quantile(x, q)
-		mask_l = x < x_l
-		mask_u = x > x_u
-		c[mask_l] = -1
-		c[mask_u] =  1
-		x[mask_l] = x_l
-		x[mask_u] = x_u
-		return x, c
-	else:
-		raise ValueError("'where' parameter not correctly defined")
+    c = np.zeros_like(x)
+    x = np.copy(x)
[…]
