### 2021-07-08 f219c50
Completed Gumbel doc page. Added Gauss and Galton as options for Normal and LogNormal
