+++ surpyval/parametric/gumbel.py
@@ -29,32 +29,284 @@ class Gumbel_(ParametricFitter):
 		return para.Gumbel.fit(x, c, n, how='MPP').params
 
 	def sf(self, x, mu, sigma):
+		r"""
+
+		Surival (or Reliability) function for the Gumbel Distribution:
+
+		.. math::
+			R(x) = 1 - e^{e^{-(x - \mu)/\sigma}}
+
+		Parameters
+		----------
+
+		x : numpy array or scalar
+			The values of the random variables at which the survival function will be calculated 
+		mu : numpy array like or scalar
+			The location parameter of the distribution
+		sigma : numpy array like or scalar
+			The scale parameter of the distribution
+
+		Returns
+		-------
+
+		sf : scalar or numpy array 
+			The scalar value of the survival function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the survival function at each corresponding value in the input array.
+
+
+		Examples
+		--------
+		>>> import numpy as np
+		>>> from surpyval import Gumbel
+		>>> x = np.array([1, 2, 3, 4, 5])
+		>>> Gumbel.sf(x, 3, 2)
+		array([0.69220063, 0.54523921, 0.36787944, 0.19229565, 0.06598804])
+		"""
 		return np.exp(-np.exp((x - mu)/sigma))
 
 	def cs(self, x, X, mu, sigma):
 		return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
 
 	def ff(self, x, mu, sigma):
+		r"""
+
+		CDF (or Failure) function for the Gumbel Distribution:
+
+		.. math::
+			F(x) = e^{e^{-(x - \mu)/\sigma}}
+
+		Parameters
+		----------
+
+		x : numpy array or scalar
+			The values of the random variables at which the survival function will be calculated 
+		mu : numpy array like or scalar
+			The location parameter of the distribution
+		sigma : numpy array like or scalar
+			The scale parameter of the distribution
+
+		Returns
+		-------
+
+		ff : scalar or numpy array 
+			The scalar value of the failure function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the failure function at each corresponding value in the input array.
+
+
+		Examples
+		--------
+		>>> import numpy as np
+		>>> from surpyval import Gumbel
+		>>> x = np.array([1, 2, 3, 4, 5])
+		>>> Gumbel.ff(x, 3, 2)
+		array([0.30779937, 0.45476079, 0.63212056, 0.80770435, 0.93401196])
+		"""
 		return 1 - np.exp(-np.exp((x - mu)/sigma))
 
 	def df(self, x, mu, sigma):
+		r"""
+
+		Density function (pdf) for the Gumbel Distribution:
+
+		.. math::
+			f(x) = \frac{1}{\sigma}e^{\left (\frac{x - \mu}{\sigma} - e^{\frac{x-\mu}{\sigma}} \right)}
+
+		Parameters
+		----------
+
+		x : numpy array or scalar
+			The values of the random variables at which the survival function will be calculated 
+		mu : numpy array like or scalar
+			The location parameter of the distribution
+		sigma : numpy array like or scalar
+			The scale parameter of the distribution
+
+		Returns
+		-------
+
+		df : scalar or numpy array 
[…]
