+++ docs/weibull.rst
@@ -1,0 +1,7 @@
+Weibull
+=======
+
+.. automodule:: surpyval.parametric.weibull
+   :members:
+   :undoc-members:
+   :inherited-members: fit
