+++ docs/loglogistic.rst
@@ -1,0 +1,7 @@
+LogLogistic
+===========
+
+.. automodule:: surpyval.parametric.loglogistic
+   :members:
+   :undoc-members:
+   :inherited-members: fit
