+++ docs/lognormal.rst
@@ -1,0 +1,7 @@
+LogNormal
+=========
+
+.. automodule:: surpyval.parametric.lognormal
+   :members:
+   :undoc-members:
+   :inherited-members: fit
