+++ docs/uniform.rst
@@ -1,0 +1,7 @@
+Uniform
+=======
+
+.. automodule:: surpyval.parametric.uniform
+   :members:
+   :undoc-members:
+   :inherited-members: fit
