+++ docs/logistic.rst
@@ -1,0 +1,7 @@
+Logistic
+========
+
+.. automodule:: surpyval.parametric.logistic
+   :members:
+   :undoc-members:
+   :inherited-members: fit
