### 2021-07-08 c755e02
Completed Exponential docs
