+++ surpyval/parametric/exponential.py
@@ -12,115 +12,400 @@ from .fitters.mpp import mpp
 import warnings
 
 class Exponential_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 1
-		self.bounds = ((0, None),)
-		self.support = (0, np.inf)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.05, 0.4, 0.6, 0.7, 0.8, 0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['lambda']
-		self.param_map = {
-			'lambda' : 0,
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 1
+        self.bounds = ((0, None),)
+        self.support = (0, np.inf)
+        self.plot_x_scale = 'linear'
+        self.y_ticks = [0.05, 0.4, 0.6, 0.7, 0.8, 0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['lambda']
+        self.param_map = {
+            'lambda' : 0,
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, offset=False):
-		x, c, n = surpyval.xcn_handler(x, c, n)
-		c = (c == 0).astype(np.int64)
-		rate = (n * c).sum()/x.sum()
-		if offset:
-			return np.min(x) - (np.max(x) - np.min(x))/10., rate
-		else:
-			return np.array([rate])
+    def _parameter_initialiser(self, x, c=None, n=None, offset=False):
+        x, c, n = surpyval.xcn_handler(x, c, n)
+        c = (c == 0).astype(np.int64)
+        rate = (n * c).sum()/x.sum()
+        if offset:
+            return np.min(x) - (np.max(x) - np.min(x))/10., rate
+        else:
+            return np.array([rate])
 
-	def sf(self, x, failure_rate):
-		return np.exp(-failure_rate * x)
+    def sf(self, x, failure_rate):
+        r"""
 
-	def cs(self, x, X, failure_rate):
-		# The exponential distribution is memoryless so of course it is the same as the survival function
-		return self.sf(x, failure_rate)
+        Surival (or Reliability) function for the Exponential Distribution:
 
-	def ff(self, x, failure_rate):
-		return 1 - np.exp(-failure_rate * x)
+        .. math::
+            R(x) = e^{-\lambda x}
 
-	def df(self, x, failure_rate):
-		return failure_rate * np.exp(-failure_rate * x)
+        Parameters
+        ----------
 
-	def hf(self, x, failure_rate):
-		return failure_rate
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        failure_rate : numpy array or scalar
+            The scale parameter for the Exponential distribution
 
-	def Hf(self, x, failure_rate):
-		return failure_rate * x
+        Returns
+        -------
 
-	def qf(self, p, failure_rate):
-		return -np.log(p)/failure_rate
+        sf : scalar or numpy array 
+            The scalar value of the survival function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the survival function at each corresponding value in the input array.
 
-	def mean(self, failure_rate):
-		return 1. / failure_rate
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Exponential
+        >>> x = np.array([1, 2, 3, 4, 5])
+        >>> Exponential.sf(x, 3)
+        array([4.97870684e-02, 2.47875218e-03, 1.23409804e-04, 6.14421235e-06,
+               3.05902321e-07])
+        """
+        return np.exp(-failure_rate * x)
 
-	def moment(self, n, failure_rate):
-		return factorial(n) / (failure_rate ** n)
+    def cs(self, x, X, failure_rate):
[…]
