+++ surpyval/parametric/lognormal.py
@@ -51,6 +51,9 @@ class LogNormal_(ParametricFitter):
 	def mean(self, mu, sigma):
 		return np.exp(mu + (sigma**2)/2)
 
+	def moment(self, n, mu, sigma):
+		return np.exp(n*mu + (n**2 * sigma**2)/2)
+
 	def random(self, size, mu, sigma):
 		return np.exp(para.Normal.random(size, mu, sigma))
 
@@ -87,6 +90,11 @@ class LogNormal_(ParametricFitter):
 	def R_cb(self, x, mu, sigma, cv_matrix, cb=0.05):
 		t = np.log(x)
 		return para.Normal.sf(self.z_cb(t, mu, sigma, cv_matrix, cb=0.05), 0, 1).T
+
+	def _mom(self, x):
+		norm_mod = para.Normal.fit(np.log(x), how='MOM')
+		mu, sigma = norm_mod.params
+		return mu, sigma
 
 LogNormal = LogNormal_('LogNormal')
 Galton = LogNormal_('Galton')
