+++ surpyval/tests/test_fit.py
@@ -29,6 +29,12 @@ def generate_mpp_test_cases():
             yield dist, bounds, rr
 
 
+def generate_mom_test_cases():
+    for idx, dist in enumerate(DISTS):
+        bounds = parameter_sample_bounds[idx]
+        yield dist, bounds
+
+
 def idfunc(x):
     if type(x) is tuple:
         return 'bounds'
@@ -150,3 +156,45 @@ def test_mpp(dist, bounds, rr):
             break
     else:
         raise AssertionError('MPP fit not very good in %s\n' % dist.name)
+
+
+@pytest.mark.parametrize("dist,bounds",
+                         generate_mom_test_cases(),
+                         ids=idfunc)
+def test_mom(dist, bounds):
+    for n in FIT_SIZES:
+        test_params = []
+        tol = 0.1
+        for b in bounds:
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
+        x = dist.random(n, *test_params)
+        model = dist.fit(x=x, how='MOM')
+        fitted_params = np.array(model.params)
+        max_params = np.max([fitted_params, test_params], axis=0)
+        diff = np.abs(fitted_params - test_params) / max_params
+        if (diff < tol).all():
+            break
+    else:
+        raise AssertionError('MOM fit not very good in %s\n' % dist.name)
+
+
+@pytest.mark.parametrize("dist,bounds",
+                         generate_mom_test_cases(),
+                         ids=idfunc)
+def test_mps(dist, bounds):
+    for n in FIT_SIZES:
+        test_params = []
+        tol = 0.05
+        for b in bounds:
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
+        x = dist.random(n, *test_params)
+        model = dist.fit(x=x, how='MPS')
+        fitted_params = np.array(model.params)
+        max_params = np.max([fitted_params, test_params], axis=0)
+        diff = np.abs(fitted_params - test_params) / max_params
+        if (diff < tol).all():
+            break
+    else:
+        raise AssertionError('MPS fit not very good in %s\n' % dist.name)
