+++ surpyval/parametric/gamma.py
@@ -19,159 +19,328 @@ from surpyval import nonparametric as nonp
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Gamma_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((0, None), (0, None),)
-		self.support = (0, np.inf)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta']
-		self.param_map = {
-			'alpha' : 0,
-			'beta'  : 1
-		}
+    r"""
 
-	def _parameter_initialiser(self, x, c=None, n=None, offset=False):
-		# These equations are truly magical
-		if offset:
-			s = np.log(x.sum()/len(x)) - np.log(x).sum()/len(x)
-			alpha = (3 - s + np.sqrt((s - 3)**2 + 24*s)) / (12*s)
-			beta = x.sum()/(len(x)*alpha)
-			return alpha, 1./beta, np.min(x) - 1
-		else:
-			s = np.log(x.sum()/len(x)) - np.log(x).sum()/len(x)
-			alpha = (3 - s + np.sqrt((s - 3)**2 + 24*s)) / (12*s)
-			beta = x.sum()/(len(x)*alpha)
-			return alpha, 1./beta
+    Class used to generate the Gamma class.
 
-	def sf(self, x, alpha, beta):
-		return 1 - self.ff(x, alpha, beta)
+    .. code:: python
 
-	def cs(self, x, X, alpha, beta):
-		return self.sf(x + X, alpha, beta) / self.sf(X, alpha, beta)
+        from surpyval import Gamma
 
-	def ff(self, x, alpha, beta):
-		return agammainc(alpha, beta * x)
+    """
+    def __init__(self, name):
+        self.name = name
+        self.k = 2
+        self.bounds = ((0, None), (0, None),)
+        self.support = (0, np.inf)
+        self.plot_x_scale = 'linear'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['alpha', 'beta']
+        self.param_map = {
+            'alpha' : 0,
+            'beta'  : 1
+        }
 
-	def df(self, x, alpha, beta):
-		return ((beta ** alpha) * x ** (alpha - 1) * np.exp(-(x * beta)) / (agamma(alpha)))
+    def _parameter_initialiser(self, x, c=None, n=None, offset=False):
+        # These equations are truly magical
+        if offset:
+            s = np.log(x.sum()/len(x)) - np.log(x).sum()/len(x)
+            alpha = (3 - s + np.sqrt((s - 3)**2 + 24*s)) / (12*s)
+            beta = x.sum()/(len(x)*alpha)
+            return alpha, 1./beta, np.min(x) - 1
+        else:
+            s = np.log(x.sum()/len(x)) - np.log(x).sum()/len(x)
+            alpha = (3 - s + np.sqrt((s - 3)**2 + 24*s)) / (12*s)
[…]
