+++ surpyval/parametric/fitters/mom.py
@@ -11,19 +11,23 @@ def mom(dist, x, n, init, offset):
 	"""
 	x_ = np.repeat(x, n)
 
+	if hasattr(dist, '_mom'):
+		return {'params' : dist._mom(x)}
+
 	if offset:
 		k = dist.k + 1
 		bounds = ((None, np.min(x)), *dist.bounds)
 	else:
 		k = dist.k
 		bounds = dist.bounds
+
 	moments = np.zeros(k)
 
 	for i in range(0, k):
-		moments[i] = np.sum(x_**(i+1)) / len(x_)
+		moments[i] = (x_**(i+1)).mean()
 
 	fun = lambda params : np.sum((moments - dist.mom_moment_gen(*params, offset=offset))**2)
-	res = minimize(fun, init, bounds=bounds)
+	res = minimize(fun, init, bounds=bounds, tol=1e-15)
 
 	results = {}
 	if offset:
