+++ docs/exponential.rst
@@ -1,6 +1,7 @@
-Exponential Distribution
-========================
+Exponential
+===========
 
 .. autoclass:: surpyval.parametric.exponential.Exponential_
    :members:
    :inherited-members: fit
+   :show-inheritance:

[…]
