### 2021-07-09 6740ec6
Deleted old mom method
