### 2021-07-09 d65bdec
Created MSE test. All passing
