+++ surpyval/parametric/fitters/mse.py
@@ -3,7 +3,7 @@ import autograd.numpy as np
 from surpyval import nonparametric as nonp
 from scipy.optimize import minimize
 
-def mse(dist, x, c, n, init):
+def mse(dist, x, c, n, t, init, const, trans, inv_fs, fixed_idx, offset):
 	"""
 	MSE: Mean Square Error
 	This is simply fitting the curve to the best estimate from a non-parametric estimate.
@@ -13,14 +13,20 @@ def mse(dist, x, c, n, init):
 	a the CDF sigmoid to the nonparametric estimate.
 	"""
 
-	x_, r, d, R = nonp.turnbull(x, c, n, estimator='Nelson-Aalen')
+	# Need to make the Turnbull estimate much much much faster (and less memory before I unlock this)
+	# x_, r, d, R = nonp.turnbull(x, c, n, t, estimator='Nelson-Aalen')
+	x_, r, d, R = nonp.nelson_aalen(x, c, n)
 
 	F = 1 - R
 	mask = np.isfinite(x_)
 	F  = F[mask]
 	x_ = x_[mask]
 	
-	fun = lambda params : np.sum(((dist.ff(x_, *params)) - F)**2)
-	res = minimize(fun, init, bounds=dist.bounds)
+	fun = lambda params : np.sum(((dist.ff(x_, *inv_fs(const(params)))) - F)**2)
+	res = minimize(fun, init)
 
-	return res
+	results = {}
+	results['res'] = res
+	results['params'] = inv_fs(const(res.x))
+
+	return results
