### 2021-07-09 66846bc
Added Beta to test_fit. Fixed bug found in Beta mom fitting
