### 2021-07-09 b6ae591
Probability plot for beta too unstable. Will not implement
