+++ surpyval/parametric/beta.py
@@ -10,6 +10,9 @@ from autograd_gamma import gammainc as agammainc
 from autograd.scipy.special import beta as abeta
 from autograd_gamma import betainc as abetainc
 from scipy.special import ndtri as z
+from scipy.linalg import lstsq
+
+import warnings
 
 from scipy.optimize import minimize
 from scipy.stats import pearsonr
@@ -20,111 +23,112 @@ from surpyval import nonparametric as nonp
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Beta_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((0, None), (0, None),)
-		self.support = (0, 1)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta']
-		self.param_map = {
-			'alpha' : 0,
-			'beta'  : 1
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 2
+        self.bounds = ((0, None), (0, None),)
+        self.support = (0, 1)
+        self.plot_x_scale = 'linear'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['alpha', 'beta']
+        self.param_map = {
+            'alpha' : 0,
+            'beta'  : 1
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None):
-		if (c is not None) & ((c == 0).all()):
-			x = np.repeat(x, n)
-			p = self._mom(x)
-		else:
-			p = 1., 1.
-		return p
+    def _parameter_initialiser(self, x, c=None, n=None):
+        if (c is not None) & ((c == 0).all()):
+            x = np.repeat(x, n)
+            p = self._mom(x)
+        else:
+            p = 1., 1.
+        return p
 
-	def sf(self, x, alpha, beta):
-		return 1 - self.ff(x, alpha, beta)
+    def sf(self, x, alpha, beta):
+        return 1 - self.ff(x, alpha, beta)
 
-	def cs(self, x, X, alpha, beta):
-		return self.sf(x + X, alpha, beta) / self.sf(X, alpha, beta)
+    def cs(self, x, X, alpha, beta):
+        return self.sf(x + X, alpha, beta) / self.sf(X, alpha, beta)
 
-	def ff(self, x, alpha, beta):
-		return abetainc(alpha, beta, x)
+    def ff(self, x, alpha, beta):
+        return abetainc(alpha, beta, x)
 
-	def df(self, x, alpha, beta):
-		return (x**(alpha - 1) * (1 - x)**(beta - 1)) / abeta(alpha, beta)
+    def df(self, x, alpha, beta):
+        return (x**(alpha - 1) * (1 - x)**(beta - 1)) / abeta(alpha, beta)
 
-	def hf(self, x, alpha, beta):
-		return self.df(x, alpha, beta) / self.sf(x, alpha, beta)
+    def hf(self, x, alpha, beta):
[…]
