### 2021-07-09 6f137cc
Completed docs for fit method
