+++ surpyval/parametric/parametric_fitter.py
@@ -25,18 +25,6 @@ from .fitters import bounds_convert, fix_idx_and_function
 PARA_METHODS = ['MPP', 'MLE', 'MPS', 'MSE', 'MOM']
 
 class ParametricFitter():
-	# def parameter_initialiser_(self, x, c, n, offset=False):
-	# 	# Change this to simply invoke MPP
-	# 	x, c, n, R = nonp.turnbull(x, c, n, estimator='Turnbull')
-	# 	F = 1 - R
-	# 	F = F[np.isfinite(x)]
-	# 	x = x[np.isfinite(x)]
-	# 	# Need to add offset step.
-	# 	init = np.ones(self.k)
-	# 	fun = lambda params : np.sum(((self.ff(x, *params)) - F)**2)
-	# 	res = minimize(fun, init, bounds=self.bounds)
-	# 	return res.x
-
 	def like(self, x, c, n, *params):
 		like = np.zeros_like(x).astype(surpyval.NUM)
 		like = np.where(c ==  0, self.df(x, *params), like)
@@ -134,12 +122,118 @@ class ParametricFitter():
 			moments[i] = self._moment(n, *params, offset=offset)
 		return moments
 
-	def fit(self, x, c=None, n=None, t=None, how='MLE', **kwargs):
-		# Check inputs
-		offset = kwargs.pop('offset', False)
-		fixed = kwargs.pop('fixed', None)
-		tl = kwargs.pop('tl', None)
-		tr = kwargs.pop('tr', None)
+	def fit(self, x=None, c=None, n=None, t=None, how='MLE',
+			offset=False, tl=None, tr=None, xl=None, xr=None,
+			fixed=None, heuristic='Nelson-Aalen', init=[], rr='y'):
+
+		r"""
+
+		The central feature to SurPyval's capability. This function aimed to have an API to mimic the 
+		simplicity of the scipy API. That is, to use a simple :code:`fit()` call, with as many or as few
+		parameters as is needed.
+
+		Parameters
+		----------
+
+		x : array like, optional
+			Array of observations of the random variables. If x is :code:`None`, xl and xr must be provided.
+		c : array like, optional
+			Array of censoring flag. -1 is left censored, 0 is observed, 1 is right censored, and 2 is intervally
+			censored. If not provided will assume all values are observed.
+		n : array like, optional
+			Array of counts for each x. If data is proivded as counts, then this can be provided. If :code:`None`
+			will assume each observation is 1.
+		t : 2D-array like, optional
+			2D array like of the left and right values at which the respective observation was truncated. If
+			not provided it assumes that no truncation occurs.
+		how : {'MLE', 'MPP', 'MOM', 'MSE', 'MPS'}, optional
+			Method to estimate parameters, these are:
+
+				- MLE : Maximum Likelihood Estimation
+				- MPP : Method of Probability Plotting
+				- MOM : Method of Moments
+				- MSE : Mean Square Error
+				- MPS : Maximum Product Spacing
+
+		offset : boolean, optional
+			If :code:`True` finds the shifted distribution. If not provided assumes not a shifted distribution.
+			Only works with distributions that are supported on the half-real line.
+
+		tl : array like or scalar, optional
+			Values of left truncation for observations. If it is a scalar value assumes each observation is
+			left truncated at the value. If an array, it is the respective 'late entry' of the observation
+
+		tr : array like or scalar, optional
+			Values of right truncation for observations. If it is a scalar value assumes each observation is
[…]
