### 2021-07-09 5c1abf7
flake8 standard compliance work
