+++ surpyval/parametric/fitters/mle.py
@@ -1,113 +1,109 @@
 import autograd.numpy as np
 from autograd import jacobian, hessian
 from autograd.numpy.linalg import inv
-from autograd.numpy.linalg import pinv
 import copy
-
 from scipy.optimize import minimize
-from scipy.optimize import approx_fprime
-
-import surpyval
-import warnings
 import sys
 
-def _create_censor_flags(x_mle, gamma, c, dist):
-	if 2 in c:
-		l_flag = x_mle[:, 0] <= np.min([gamma, dist.support[0]])
-		r_flag = x_mle[:, 1] >= dist.support[1]
-		mask = np.vstack([l_flag, r_flag]).T
-		inf_c_flags = (mask).astype(int)
-		x_mle[(c == 2).reshape(-1, 1) & mask] = 1
-	else:
-		inf_c_flags = np.zeros_like(x_mle)
 
-	return inf_c_flags, x_mle
+def _create_censor_flags(x_mle, gamma, c, dist):
+    if 2 in c:
+        l_flag = x_mle[:, 0] <= np.min([gamma, dist.support[0]])
+        r_flag = x_mle[:, 1] >= dist.support[1]
+        mask = np.vstack([l_flag, r_flag]).T
+        inf_c_flags = (mask).astype(int)
+        x_mle[(c == 2).reshape(-1, 1) & mask] = 1
+    else:
+        inf_c_flags = np.zeros_like(x_mle)
+
+    return inf_c_flags, x_mle
+
 
 def mle(dist, x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset):
-		"""
-		Maximum Likelihood Estimation (MLE)
+    """
+    Maximum Likelihood Estimation (MLE)
 
-		"""
-		if hasattr(dist, 'mle'):
-			return dist.mle(x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset)
+    """
+    if hasattr(dist, 'mle'):
+        return dist.mle(x, c, n, t, const, trans,
+                        inv_fs, init, fixed_idx, offset)
 
-		old_err_state = np.seterr(invalid='raise', divide='raise', over='ignore', under='ignore')
+    old_err_state = np.seterr(invalid='raise',
+                              divide='raise',
+                              over='ignore',
+                              under='ignore')
 
-		results = {}
+    results = {}
 
-		# Need to flag entries where truncation is inf or -inf so that the autograd doesn't fail. Because
-		# autograd fails if it encounters any inf, nan, -inf etc even if they don't affect the gradient
-		# Very important to the autograd! And it worked, yay!
-		t_flags = np.ones_like(t)
-		t_mle = copy.copy(t)
-		# Create flags to indicate where the truncation values are infinite
-		t_flags[:, 0] = np.where(np.isfinite(t[:, 0]), 1, 0)
-		t_flags[:, 1] = np.where(np.isfinite(t[:, 1]), 1, 0)
-		# Convert the infinite values to a finite value to ensure the autodiff functions don't fail
-		t_mle[:, 0] = np.where(t_flags[:, 0] == 1, t[:, 0], 1)
-		t_mle[:, 1] = np.where(t_flags[:, 1] == 1, t[:, 1], 1)
+    # Need to flag entries where truncation is inf or -inf so that the autograd doesn't fail. Because
+    # autograd fails if it encounters any inf, nan, -inf etc even if they don't affect the gradient
+    # Very important to the autograd! And it worked, yay!
+    t_flags = np.ones_like(t)
+    t_mle = copy.copy(t)
+    # Create flags to indicate where the truncation values are infinite
[…]
