### 2021-07-09 a9f45ae
Fixed offset bug in Weibull. Added docstring for fit_from_df
