### 2021-07-09 7d4c999
Fixed bug in weibull.py
