### 2021-07-09 7c4e8cb
removed 'surpyval' import in exponential
