### 2021-07-09 3bba65e
removed import surpyval where unnecessary
