### 2021-07-09 bfb41fd
Added additional size for test to address unifrom fit. Removed double Weibull reference in docs
