+++ surpyval/tests/test_fit.py
@@ -12,7 +12,7 @@ parameter_sample_bounds = [((1, 20), (0.5, 5)),
                            ((1, 100), (0.5, 20)),
                            ((1, 100), (1, 100)),
                            ((.1, 30), (.1, 30)),]
-FIT_SIZES = [5000, 10000, 20000, 50000, 100000]
+FIT_SIZES = [5000, 10000, 20000, 50000, 100000, 1000000]
 
 
 def generate_mle_test_cases():
