+++ README.md
@@ -1,6 +1,6 @@
 # SurPyval - Survival Analysis in Python
 
-[![PyPI - Version](https://img.shields.io/badge/pypi-v0.4.0-success)](https://pypi.org/project/surpyval/)
+[![PyPI - Version](https://img.shields.io/badge/pypi-v0.5.0-success)](https://pypi.org/project/surpyval/)
 ![PyPI - Python Version](https://img.shields.io/badge/python-3.6%20%7C%203.7%20%7C%203.8%20%7C%203.9-blue)
 [![Documentation Status](https://readthedocs.org/projects/surpyval/badge/?version=latest)](https://surpyval.readthedocs.io/en/latest/?badge=latest)
 
