+++ surpyval/tests/test_fit.py
@@ -176,8 +176,8 @@ def test_mpp(dist, bounds, rr):
 def test_mom(dist, bounds):
     for n in FIT_SIZES:
         test_params = []
-        # 1% accuracy!!
-        tol = 0.01
+        # 5% accuracy!!
+        tol = 0.05
         for b in bounds:
             test_params.append(np.random.uniform(*b))
         test_params = np.array(test_params)
