### 2021-07-11 699a8e7
Changed distributions to autoclass instead of module. Updated docstrings in gamma and exponential
