+++ docs/gumbel.rst
@@ -3,4 +3,4 @@ Gumbel
 
 .. autoclass:: surpyval.parametric.gumbel.Gumbel_
    :members:
-   :inherited-members:
+   :show-inheritance:
