+++ docs/exponential.rst
@@ -3,5 +3,4 @@ Exponential
 
 .. autoclass:: surpyval.parametric.exponential.Exponential_
    :members:
-   :inherited-members: fit
    :show-inheritance:
