### 2021-07-12 fe53e29
More docstring updates for normal, gumbel, exponential, parametric_fitter
