+++ surpyval/parametric/normal.py
@@ -9,6 +9,15 @@ from surpyval import parametric as para
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Normal_(ParametricFitter):
+	r"""
+
+    Class used to generate the Normal (Gauss) class.
+
+    .. code:: python
+
+        from surpyval import Normal
+
+    """
 	def __init__(self, name):
 		self.name = name
 		self.k = 2
@@ -25,38 +34,341 @@ class Normal_(ParametricFitter):
 
 	def _parameter_initialiser(self, x, c=None, n=None, t=None):
 		return para.Normal.fit(x, c, n, t, how='MPP').params
-		# x, c, n = surpyval.xcn_handler(x, c, n)
-		# flag = (c == 0).astype(int)
-		# return x.sum() / (n * flag).sum(), np.std(x)
 
 	def sf(self, x, mu, sigma):
+		r"""
+
+        Surival (or Reliability) function for the Normal Distribution:
+
+        .. math::
+            R(x) = 1 - \Phi \left( \frac{x - \mu}{\sigma} \right )
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        mu : numpy array or scalar
+            The location parameter for the Normal distribution
+        sigma : numpy array or scalar
+            The scale parameter for the Normal distribution
+
+        Returns
+        -------
+
+        sf : scalar or numpy array 
+            The value(s) of the survival function at x.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Normal
+        >>> x = np.array([1, 2, 3, 4, 5])
+        >>> Normal.sf(x, 3, 4)
+        array([0.69146246, 0.59870633, 0.5       , 0.40129367, 0.30853754])
+        """
 		return norm.sf(x, mu, sigma)
 
 	def cs(self, x, X, mu, sigma):
+		r"""
+
+        Conditional survival function for the Normal Distribution:
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The value(s) at which the function will be calculated
+        X : numpy array or scalar
+            The value(s) at which each value(s) in x was known to have survived
+        mu : numpy array or scalar
+            The location parameter for the Normal distribution
+        sigma : numpy array or scalar
+            The scale parameter for the Normal distribution
+
+        Returns
+        -------
+
+        cs : scalar or numpy array 
+            the conditional survival probability at x
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Normal
+        >>> x = np.array([1, 2, 3, 4, 5])
+        >>> Normal.cs(x, 5, 3, 4)
+        array([0.73452116, 0.51421702, 0.34242113, 0.2165286 , 0.1298356 ])
+        """
 		return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
 
 	def ff(self, x, mu, sigma):
+		r"""
+
+        CDF (or unreliability or failure) function for the Normal Distribution:
+
+        .. math::
+            F(x) = \Phi \left( \frac{x - \mu}{\sigma} \right )
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        mu : numpy array or scalar
+            The location parameter for the Normal distribution
+        sigma : numpy array or scalar
+            The scale parameter for the Normal distribution
+
+        Returns
+        -------
+
+        ff : scalar or numpy array 
+            The value(s) of the failure function at x.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Normal
[…]
