### 2021-07-12 70f8397
Added mixture models, and fixed a bug, to the docs
