### 2021-07-12 bab4127
fixed flake8 error in parametric_fitter
