### 2021-07-12 768595d
Doc updates to include LFP and MSE parametric estimation.
