### 2021-07-12 #3484 @whedon
[REVIEW]: Survival: Survival Analysis in Python
**Submitting author:** @derrynknife (<a href="http://orcid.org/0000-0002-6636-6318">Derryn Knife</a>)
**Repository:** <a href="https://github.com/derrynknife/SurPyval" target ="_blank">https://github.com/derrynknife/SurPyval</a>
**Version:** 0.10.0
