### 2021-07-12 #3484 @dfm
@derrynknife, @CamDavidsonPilon, @MatthewReid854 – This is the review thread for the paper. All of our communications will happen here from now on. Thanks again for agreeing to participate!
