### 2021-07-12 derrynknife/SurPyval#3 @CamDavidsonPilon
[JOSS review] License clarification
Hi @derrynknife, I'm one of the reviewers for your JOSS article. I can see the project is still being worked on, so I'll hold back some of my review for a few weeks until it stabilizes more. I'll create some minor issues for now, though.

In the [MIT license](https://github.com/derrynknife/SurPyval/blob/master/LICENSE) in this repo, you have `The Python Packaging Authority` as the holder - I suspect you instead want yourself there, unless you indeed do want the Python Packaging Authority to have the copyright?

https://github.com/openjournals/joss-reviews/issues/3484
