### 2021-07-12 derrynknife/SurPyval#4 @CamDavidsonPilon
[JOSS Review] Parametric form in docs 
One of the annoying things about statistics is all the different ways one can parameterize a distribution. For absolute clarity to users, I think it's important to have the parametric form used in the docs. It's been done for some [distributions](https://surpyval.readthedocs.io/en/latest/exponential.html), but needs to be there for all (example [missing](https://surpyval.readthedocs.io/en/latest/lognormal.html).

https://github.com/openjournals/joss-reviews/issues/3484
