+++ surpyval/parametric/beta.py
@@ -48,33 +48,339 @@ class Beta_(ParametricFitter):
         return p
 
     def sf(self, x, alpha, beta):
+        r"""
+
+        Survival (or reliability) function for the Beta Distribution:
+
+        .. math::
+            R(x) = 1 - \int_{0}^{x}t^{\alpha-1}\left (1 - t \right )^{\beta -1}dt
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        alpha : numpy array or scalar
+            One shape parameter for the Beta distribution
+        beta : numpy array or scalar
+            The scale parameter for the Beta distribution
+
+        Returns
+        -------
+
+        df : scalar or numpy array 
+            The value(s) of the reliability function at x.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Beta
+        >>> x = np.array([.1, .2, .3, .4, .5])
+        >>> Beta.sf(x, 3, 4)
+        array([0.98415, 0.90112, 0.74431, 0.54432, 0.34375])
+        """
         return 1 - self.ff(x, alpha, beta)
 
     def cs(self, x, X, alpha, beta):
+        r"""
+
+        Conditional survival (or reliability) function for the Beta Distribution:
+
+        .. math::
+            R(x, X) = \frac{R(x + X)}{R(X)}
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        X : numpy array or scalar
+            The value(s) at which each value(s) in x was known to have survived
+        alpha : numpy array or scalar
+            One shape parameter for the Beta distribution
+        beta : numpy array or scalar
+            The scale parameter for the Beta distribution
+
+        Returns
+        -------
+
+        df : scalar or numpy array 
+            The value(s) of the conditional survival function at x.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Beta
+        >>> x = np.array([.1, .2, .3, .4, .5])
+        >>> Beta.cs(x, 0.4, 3, 4)
+        array([0.6315219 , 0.32921811, 0.12946429, 0.03115814, 0.00233319])
+        """
         return self.sf(x + X, alpha, beta) / self.sf(X, alpha, beta)
 
     def ff(self, x, alpha, beta):
+        r"""
+
+        Failure (CDF or unreliability) function for the Beta Distribution:
+
+        .. math::
+            F(x) = \int_{0}^{x}t^{\alpha-1}\left (1 - t \right )^{\beta -1}dt
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        alpha : numpy array or scalar
+            One shape parameter for the Beta distribution
+        beta : numpy array or scalar
+            The scale parameter for the Beta distribution
+
+        Returns
+        -------
+
+        df : scalar or numpy array 
+            The value(s) of the failure function at x.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from surpyval import Beta
+        >>> x = np.array([.1, .2, .3, .4, .5])
+        >>> Beta.ff(x, 3, 4)
+        array([0.01585, 0.09888, 0.25569, 0.45568, 0.65625])
+        """
         return abetainc(alpha, beta, x)
 
     def df(self, x, alpha, beta):
+        r"""
+
+        Density function for the Beta Distribution:
+
+        .. math::
+            f(x) = \frac{x^{\alpha-1}\left(1 - x \right )^{\beta-1}}{B \left ( \alpha , \beta \right )}
+
+        Parameters
+        ----------
+
+        x : numpy array or scalar
[…]
