### 2021-07-13 7a63fc6
Updated doc strings for logistic
