+++ surpyval/parametric/logistic.py
@@ -6,81 +6,328 @@ from surpyval import xcn_handler
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Logistic_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((None, None), (0, None),)
-		self.support = (-np.inf, np.inf)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['mu', 'sigma']
-		self.param_map = {
-			'mu'    : 0,
-			'sigma' : 1
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 2
+        self.bounds = ((None, None), (0, None),)
+        self.support = (-np.inf, np.inf)
+        self.plot_x_scale = 'linear'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['mu', 'sigma']
+        self.param_map = {
+            'mu'    : 0,
+            'sigma' : 1
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
-		return self.fit(x, c, n, t, how='MPP').params
-		x, c, n = xcn_handler(x, c, n)
-		flag = (c == 0).astype(int)
-		if offset:
-			return x.sum() / (n * flag).sum(), 1., 1.
-		else:
-			return x.sum() / (n * flag).sum(), 1.
+    def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
+        return self.fit(x, c, n, t, how='MPP').params
+        x, c, n = xcn_handler(x, c, n)
+        flag = (c == 0).astype(int)
+        if offset:
+            return x.sum() / (n * flag).sum(), 1., 1.
+        else:
+            return x.sum() / (n * flag).sum(), 1.
 
-	def sf(self, x, mu, sigma):
-		return 1 - self.ff(x, mu, sigma)
+    def sf(self, x, mu, sigma):
+        r"""
 
-	def cs(self, x, X, mu, sigma):
-		return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
+        Survival (or reliability) function for the Logistic Distribution:
 
-	def ff(self, x, mu, sigma):
-		z = (x - mu) / sigma
-		return 1. / (1 + np.exp(-z))
+        .. math::
+            R(x) = 1 - \frac{1}{1 + e^{- \left ( x - mu \right )/ \sigma}}
 
-	def df(self, x, mu, sigma):
-		z = (x - mu) / sigma
-		return np.exp(-z) / (sigma * (1 + np.exp(-z))**2)
+        Parameters
+        ----------
 
-	def hf(self, x, mu, sigma):
-		return self.pdf(x, mu, sigma) / self.sf(x, mu, sigma)
+        x : numpy array or scalar
+            The values at which the function will be calculated 
+        mu : numpy array or scalar
+            The location parameter for the Logistic distribution
+        sigma : numpy array or scalar
+            The scale parameter for the Logistic distribution
 
-	def Hf(self, x, mu, sigma):
[…]
