### 2021-07-13 b42cff2
Updated docstrings for uniform
