+++ surpyval/parametric/normal.py
@@ -9,7 +9,7 @@ from surpyval import parametric as para
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Normal_(ParametricFitter):
-	r"""
+    r"""
 
     Class used to generate the Normal (Gauss) class.
 
@@ -18,25 +18,25 @@ class Normal_(ParametricFitter):
         from surpyval import Normal
 
     """
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((None, None), (0, None),)
-		self.support = (-np.inf, np.inf)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.001, 0.01, 0.1, 0.2, 0.3, 0.4, 
-				0.5, 0.6, 0.7, 0.8, 0.9, 0.99, 0.999]
-		self.param_names = ['mu', 'sigma']
-		self.param_map = {
-			'mu'    : 0,
-			'sigma' : 1
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 2
+        self.bounds = ((None, None), (0, None),)
+        self.support = (-np.inf, np.inf)
+        self.plot_x_scale = 'linear'
+        self.y_ticks = [0.001, 0.01, 0.1, 0.2, 0.3, 0.4, 
+                0.5, 0.6, 0.7, 0.8, 0.9, 0.99, 0.999]
+        self.param_names = ['mu', 'sigma']
+        self.param_map = {
+            'mu'    : 0,
+            'sigma' : 1
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, t=None):
-		return para.Normal.fit(x, c, n, t, how='MPP').params
+    def _parameter_initialiser(self, x, c=None, n=None, t=None):
+        return para.Normal.fit(x, c, n, t, how='MPP').params
 
-	def sf(self, x, mu, sigma):
-		r"""
+    def sf(self, x, mu, sigma):
+        r"""
 
         Surival (or Reliability) function for the Normal Distribution:
 
@@ -67,10 +67,10 @@ class Normal_(ParametricFitter):
         >>> Normal.sf(x, 3, 4)
         array([0.69146246, 0.59870633, 0.5       , 0.40129367, 0.30853754])
         """
-		return norm.sf(x, mu, sigma)
+        return norm.sf(x, mu, sigma)
 
-	def cs(self, x, X, mu, sigma):
-		r"""
+    def cs(self, x, X, mu, sigma):
+        r"""
 
         Conditional survival function for the Normal Distribution:
 
@@ -103,10 +103,10 @@ class Normal_(ParametricFitter):
         >>> Normal.cs(x, 5, 3, 4)
         array([0.73452116, 0.51421702, 0.34242113, 0.2165286 , 0.1298356 ])
         """
-		return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
+        return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
 
-	def ff(self, x, mu, sigma):
-		r"""
+    def ff(self, x, mu, sigma):
+        r"""
 
         CDF (or unreliability or failure) function for the Normal Distribution:
 
@@ -137,10 +137,10 @@ class Normal_(ParametricFitter):
         >>> Normal.ff(x, 3, 4)
         array([0.30853754, 0.40129367, 0.5       , 0.59870633, 0.69146246])
         """
-		return norm.cdf(x, mu, sigma)
+        return norm.cdf(x, mu, sigma)
 
-	def df(self, x, mu, sigma):
-		r"""
+    def df(self, x, mu, sigma):
+        r"""
 
         Density function for the Normal Distribution:
 
[…]
