+++ surpyval/parametric/loglogistic.py
@@ -9,87 +9,367 @@ from surpyval import parametric as para
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class LogLogistic_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((0, None), (0, None),)
-		self.support = (0, np.inf)
-		self.plot_x_scale = 'log'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta']
-		self.param_map = {
-			'alpha' : 0,
-			'beta' : 1
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 2
+        self.bounds = ((0, None), (0, None),)
+        self.support = (0, np.inf)
+        self.plot_x_scale = 'log'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['alpha', 'beta']
+        self.param_map = {
+            'alpha' : 0,
+            'beta' : 1
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
-		if offset:
-			# return *self.fit(x, c, n, t, how='MPP').params, 1.
-			x, c, n = xcn_handler(x, c, n)
-			flag = (c == 0).astype(int)
-			value_range = np.max(x) - np.min(x)
-			gamma_init = np.min(x) - value_range / 10
-			return gamma_init, x.sum() / (n * flag).sum(), 2., 1.
-		else:
-			return self.fit(x, c, n, t, how='MPP').params
-			# x, c, n = surpyval.xcn_handler(x, c, n)
-			# flag = (c == 0).astype(int)
-			# return x.sum() / (n * flag).sum(), 2.
+    def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
+        if offset:
+            # return *self.fit(x, c, n, t, how='MPP').params, 1.
+            x, c, n = xcn_handler(x, c, n)
+            flag = (c == 0).astype(int)
+            value_range = np.max(x) - np.min(x)
+            gamma_init = np.min(x) - value_range / 10
+            return gamma_init, x.sum() / (n * flag).sum(), 2., 1.
+        else:
+            return self.fit(x, c, n, t, how='MPP').params
+            # x, c, n = surpyval.xcn_handler(x, c, n)
+            # flag = (c == 0).astype(int)
+            # return x.sum() / (n * flag).sum(), 2.
 
-	def sf(self, x, alpha, beta):
-		return 1 - self.ff(x, alpha, beta)
+    def sf(self, x, alpha, beta):
+        r"""
 
-	def cs(self, x, X, alpha, beta):
-		return self.sf(x + X, alpha, beta) / self.sf(X, alpha, beta)
+        Survival (or reliability) function for the LogLogistic Distribution:
 
-	def ff(self, x, alpha, beta):
-		return 1. / (1 + (x/alpha)**-beta)
+        .. math::
+            R(x) = 1 - \frac{1}{1 + \left ( x /\alpha \right )^{-\beta}}
 
-	def df(self, x, alpha, beta):
[…]

[…]
