### 2021-07-13 a279065
Updated docstrings for Weibull
