+++ surpyval/parametric/weibull.py
@@ -10,151 +10,455 @@ from surpyval.parametric.parametric_fitter import ParametricFitter
 from .fitters.mpp import mpp
 
 class Weibull_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		# Set 'k', the number of parameters
-		self.k = 2
-		self.bounds = ((0, None), (0, None),)
-		self.support = (0, np.inf)
-		self.plot_x_scale = 'log'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta']
-		self.param_map = {
-			'alpha' : 0,
-			'beta' : 1
-		}
+    def __init__(self, name):
+        self.name = name
+        # Set 'k', the number of parameters
+        self.k = 2
+        self.bounds = ((0, None), (0, None),)
+        self.support = (0, np.inf)
+        self.plot_x_scale = 'log'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['alpha', 'beta']
+        self.param_map = {
+            'alpha' : 0,
+            'beta' : 1
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
-		log_x = np.log(x)
-		log_x[np.isnan(log_x)] = -np.inf
-		if offset:
-			results =  mpp(dist=self, x=x, c=c, n=n, t=t, on_d_is_0=True, offset=True)
-			return (results['gamma'], *results['params'])
-		else:
-			gumb = para.Gumbel.fit(log_x, c, n, t, how='MLE')
-			if not gumb.res.success:
-				gumb = para.Gumbel.fit(log_x, c, n, t, how='MPP')
-			mu, sigma = gumb.params
-			alpha, beta = np.exp(mu), 1. / sigma
-			if (np.isinf(alpha) | np.isnan(alpha)):
-				alpha = np.median(x)
-			if (np.isinf(beta) | np.isnan(beta)):
-				beta = 1.
-			return alpha, beta
+    def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
+        log_x = np.log(x)
+        log_x[np.isnan(log_x)] = -np.inf
+        if offset:
+            results =  mpp(dist=self, x=x, c=c, n=n, t=t, on_d_is_0=True, offset=True)
+            return (results['gamma'], *results['params'])
+        else:
+            gumb = para.Gumbel.fit(log_x, c, n, t, how='MLE')
+            if not gumb.res.success:
+                gumb = para.Gumbel.fit(log_x, c, n, t, how='MPP')
+            mu, sigma = gumb.params
+            alpha, beta = np.exp(mu), 1. / sigma
+            if (np.isinf(alpha) | np.isnan(alpha)):
+                alpha = np.median(x)
+            if (np.isinf(beta) | np.isnan(beta)):
+                beta = 1.
+            return alpha, beta
 
-	def sf(self, x, alpha, beta):
-		return np.exp(-(x / alpha)**beta)
+    def sf(self, x, alpha, beta):
+        r"""
 
[…]
