### 2021-07-13 #3484 @MatthewReid854
Overall a good article to provide an introduction to SurPyval. The following issues are intended to improve the article so please take them as constructive feedback not criticisms.

Issues identified:

- Line 5 - "scientist" should be "scientists"
- Line 5 - Starts with "Survival analysis is a tool" however the next sentence starts with "Survival analysis is a unique set of tools". Discrepancy in description of survival analysis.
- Line 9 - "in future" should be "in the future"
- Line 12 - Consider revising "insurance" to "actuarial studies" and replace "policies" with "insurance policies" on line 13
- Line 14 - Consider revising "bugged with unique problems with data" to "challenged by the availability of data".
- Line 15 - Insert comma after the word engineering
- Line 15 - plural/singular discrepancy between "components" and "it might fail"
- Line 29 - reliability needs to be referenced. Please use: https://doi.org/10.5281/zenodo.3937999
- Line 30 - States that reliability is unable to use offset values and fix parameters. Reliability includes 3 parameter distributions (with the 3rd parameter being the location or threshold parameter) for Weibull, Lognormal, Loglogistic, and Gamma distributions. Unless I am mistaken, this is the same as "offset values". Reliability enables users to fix the shape parameter for the Weibull, Lognormal, and Normal distributions. This is essential for accelerated life testing. Please revise this sentence to more accurately state the capabilities that SurPyval has which reliability does not have. I will not speak for Cam on whether lifelines can do these things.
- Line 37 - consider spelling out MLE in full as you have for all the other estimation methods
- Line 40 - I would argue that having multiple estimation methods does not make SurPyval a "much more capable package than is currently available in the Python ecosystem". We are not competing here to be the best so you do not need to sell SurPyval as "much more capable" that others. Suggested revision "This variety of estimation methods gives users of SurPyval greater flexibility in their choice of which survival analysis technique to use when fitting distributions to a data set".
- Line 45 - end the sentence with "estimation" as is used on line 50
- Line 46 - replace "capability" with "estimation" as is used on line 50
- Line 47 - MLE is missing "Estimation" when spelled out
- Line 53 - refers to Table 1 but the table below is not given a title
- Line 56 and 64 - consider replacing the word "variable" with "event" or "measurement". Variable has a special meaning in computing which makes x, c, n and t all variables.
- Line 63 - xcnt needs quotes around it
- Line 66 - please avoid claiming something is trivial as it is likely anything but trivial for most people. Consider replacing "trivial" with "possible"
- Line 66 - states "KM, NA, or FH distributions". These are non-parametric estimates not "distributions"
- Line 66 - mentions KM, NA, FH. Is it missing Turnbull in this list?
- Line 71 - NPMLE needs to be spelled out
- Lines 67 to 74 are largely a repeat of table 1. Consider removing this paragraph if it does not add value.
- Line 81 - reliability also allows fixing of shape parameters in Weibull, Lognormal and Normal distributions. Consider revising the sentence to explain whether fixing the parameters is possible for either parameter in all supported distributions.
- Line 90 - this sentence claims that SurPyval will allow the Weibull distribution's alpha parameter to be negative. This makes no sense as alpha and beta must be positive for the Weibull distribution. If you have developed some other pioneering distribution that allows alpha to be negative then consider writing a paper on that alone. Also, if you struggle with getting the optimizer to keep estimates within the positive domain, I overcame this problem in reliability using scipy's bounded optimizers (TNC, L-BFGS-B, powell, nelder-mead).
- Line 97 - claims gamma can't be estimated by other software. This statement is incorrect. See reliability.Fitters.Fit_Weibull_3P
- Line 107 - this claim needs verification. The authors of both reliability and lifelines have done a lot of work to ensure stability of optimisation. If claiming a "substantial improvement" over other packages, I would like to see evidence using a Monte Carlo simulation.
- Line 111 - When I run the example provided there is no output. Consider adding a line that produces output such as print(model). When I print the model I get "Parametric Surpyval model with Weibull distribution fitted by MLE yielding parameters [10.10167167  2.12949572]".
- Line 112 and 113 need to better introduce the code block as an example. eg. "Secondly, we can use offsets with data...." End the sentence with a colon.
- Line 114 - please hyperlink the word "documentation" to the SurPyval documentation
