### 2021-07-13 6a5634d
Fixed error in latex for docstring of df of LogLogistic
