### 2021-07-13 4e9d540
Fixed error in latex for docstring of mean of LogLogistic
