### 2021-07-13 0bc930b
Corrected errors in docstrings in uniform and weibull
