+++ paper.bib
@@ -127,3 +127,14 @@
   publisher={New York},
   doi={10.1115/1.4010337}
 }
+
+@software{matthew_reid_2021_5030847,
+  author       = {Matthew Reid},
+  title        = {MatthewReid854/reliability: v0.5.7},
+  month        = jun,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v0.5.7},
+  doi          = {10.5281/zenodo.5030847},
+  url          = {https://doi.org/10.5281/zenodo.5030847}
+}
