### 2021-07-13 fe366c6
Added ExpoWeibull to the docs
