+++ surpyval/parametric/expo_weibull.py
@@ -14,96 +14,330 @@ from surpyval.parametric.parametric_fitter import ParametricFitter
 from .fitters.mpp import mpp
 
 class ExpoWeibull_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 3
-		self.bounds = ((0, None), (0, None), (0, None),)
-		self.support = (0, np.inf)
-		self.plot_x_scale = 'log'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta', 'mu']
-		self.param_map = {
-			'alpha' : 0,
-			'beta'  : 1,
-			'mu'    : 2
-		}
+    def __init__(self, name):
+        self.name = name
+        self.k = 3
+        self.bounds = ((0, None), (0, None), (0, None),)
+        self.support = (0, np.inf)
+        self.plot_x_scale = 'log'
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
+            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
+            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
+            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.param_names = ['alpha', 'beta', 'mu']
+        self.param_map = {
+            'alpha' : 0,
+            'beta'  : 1,
+            'mu'    : 2
+        }
 
-	def _parameter_initialiser(self, x, c=None, n=None, offset=False):
-		log_x = np.log(x)
-		log_x[np.isnan(log_x)] = 0
-		gumb = para.Gumbel.fit(log_x, c, n, how='MLE')
-		if not gumb.res.success:
-			gumb = para.Gumbel.fit(log_x, c, n, how='MPP')
-		mu, sigma = gumb.params
-		alpha, beta = np.exp(mu), 1. / sigma
-		if (np.isinf(alpha) | np.isnan(alpha)):
-			alpha = np.median(x)
-		if (np.isinf(beta) | np.isnan(beta)):
-			beta = 1.
-		if offset:
-			gamma = np.min(x) - (np.max(x) - np.min(x))/10.
-			return gamma, alpha, beta, 1.
-		else:
-			return alpha, beta, 1.
-	def sf(self, x, alpha, beta, mu):
-		return 1 - np.power(1 - np.exp(-(x / alpha)**beta), mu)
+    def _parameter_initialiser(self, x, c=None, n=None, offset=False):
+        log_x = np.log(x)
+        log_x[np.isnan(log_x)] = 0
+        gumb = para.Gumbel.fit(log_x, c, n, how='MLE')
+        if not gumb.res.success:
+            gumb = para.Gumbel.fit(log_x, c, n, how='MPP')
+        mu, sigma = gumb.params
+        alpha, beta = np.exp(mu), 1. / sigma
+        if (np.isinf(alpha) | np.isnan(alpha)):
+            alpha = np.median(x)
+        if (np.isinf(beta) | np.isnan(beta)):
+            beta = 1.
+        if offset:
+            gamma = np.min(x) - (np.max(x) - np.min(x))/10.
+            return gamma, alpha, beta, 1.
+        else:
+            return alpha, beta, 1.
+    def sf(self, x, alpha, beta, mu):
+        r"""
 
[…]
