+++ docs/expoweibull.rst
@@ -1,0 +1,6 @@
+Exponentiated Weibull
+=====================
+
+.. autoclass:: surpyval.parametric.expo_weibull.ExpoWeibull_
+   :members:
+   :show-inheritance:
