+++ README.md
@@ -3,6 +3,8 @@
 [![PyPI version](https://badge.fury.io/py/surpyval.svg)](https://badge.fury.io/py/surpyval)
 ![PyPI - Python Version](https://img.shields.io/badge/python-3.6%20%7C%203.7%20%7C%203.8%20%7C%203.9-blue)
 [![Documentation Status](https://readthedocs.org/projects/surpyval/badge/?version=latest)](https://surpyval.readthedocs.io/en/latest/?badge=latest)
+Markdown: [![status](https://joss.theoj.org/papers/5a41c8e82b8327b9a3575f9c41f73ab9/status.svg)](https://joss.theoj.org/papers/5a41c8e82b8327b9a3575f9c41f73ab9)
+
 
 
 
