### 2021-07-13 2398158
Removed word markdown preceding badge
