### 2021-07-13 cb1eedd
Changed __str__ to __repr__. Closes issue 5
