### 2021-07-13 derrynknife/SurPyval#4 @derrynknife
Thanks @CamDavidsonPilon, I've updated all the docstrings for each distribution. The parameterisation should now be clear for all the distributions.

Couldn't agree more that different parameterisations can be annoying.
