### 2021-07-13 2ae8321
More minor docstring updates
