### 2021-07-14 #3484 @MatthewReid854
@derrynknife In the checklist above, we are asked to verify whether the software has automated tests and community guidelines. I am unable to tick these items off as they are not present in the library. The community guidelines are fairly easy (github has templates for these) but the automated tests are quite time consuming. Having automated tests would also help me tick off functionality from the checklist.
Derryn, do you plan on adding automated tests to SurPyval in order to get this article published? If the answer is no, then we would need to consult @dfm on whether we can tick these items off without them being present?
