### 2021-07-14 #3484 @derrynknife
There are automated tests available in the ‘surpyval/tests’ directory. This uses the ‘pytest’ module. I can add the test method to the docs if needed.

Understand the requirement for community guidelines. I can add that to the docs.
