### 2021-07-15 4c680af
Created bounds conversion with tanh to work with LFP. Works really well.
