### 2021-07-15 4cacc25
Added dist from cumulative hazard feature. Improved the bound conversions
