+++ surpyval/parametric/__init__.py
@@ -36,3 +36,5 @@ from .mixture_model import MixtureModel
 
 from .lfp import LFP
 
+from .dist_from_hazard import Dist
+
