### 2021-07-15 984ff9b
Added community guidelines
