+++ docs/Support.rst
@@ -1,0 +1,6 @@
+Support
+=======
+
+If you need help with survival analysis, please ask a question on stats.stackexchange.
+
+If you've searched the surpyval documentation for what you've been looking for and can't find it, please add as suggestion for a feature on GitHub. SurPyval is a growing tool. Or, if you need help with surpyval feel free to email Derryn at derryn.knife@gmail.com.
