+++ docs/Contributing.rst
@@ -1,0 +1,7 @@
+Conntributing
+=============
+
+If you want to contribute to SurPyval, please do! Please review the current open `feature reqeusts 
+<https://github.com/derrynknife/SurPyval/issues?q=is%3Aissue+is%3Aopen+label%3Aenhancement>`_ to see if your desired feature is in the requests. If not, please raise a new one to notify the community. We can assign you feature for you to branch and develop.
+
+SurPyval is in the process of complying with the PEP8 standard so please make all contributions as per that standard.
