### 2021-07-16 #3484 @MatthewReid854
A few more issues from my second read through:

Line 48 - Typo in "module"
Line 49 - Consider using "Method of Probability Plotting" since that's why you abbreviate to MPP
Line 77 - Typo in "observations"
Line 80 - "Minimum Spacing" should be "Minimum Product Spacing"
Line 91 - first comma (after "fixed") should be a full stop and new sentence
Line 103 - "produce" should be "producing"
Line 123 - "true" should be "True"

In order to tick off the performance and functionality items from the checklist, I want to run all the examples from your documentation. When I tried to do this, I encountered a couple of bugs for the examples on this page: https://surpyval.readthedocs.io/en/latest/SurPyval%20Models.html
It would be nice if you could include your import statements in all your examples so that users can run them in isolation.
