### 2021-07-17 d1e03c7
Corrected FH formula in docs
