### 2021-07-17 73eed4d
Fixed error in Turnbull equations
