### 2021-07-17 01c3013
Completed plotting positions and made ammendments to nonparametric docstrings
