+++ surpyval/nonparametric/turnbull.py
@@ -3,111 +3,125 @@ from surpyval import nonparametric as nonp
 from surpyval.nonparametric.nonparametric_fitter import NonParametricFitter
 
 def _na(d, r):
-	H = np.cumsum(d/r)
-	H[np.isnan(H)] = np.inf
-	R = np.exp(-H)
-	p = np.abs(np.diff(np.hstack([[1], R])))
-	return R
-	
+    H = np.cumsum(d/r)
+    H[np.isnan(H)] = np.inf
+    R = np.exp(-H)
+    p = np.abs(np.diff(np.hstack([[1], R])))
+    return R
+    
 def _km(d, r):
-	R = 1 - d/r
-	R[np.isnan(R)] = 0
-	R = np.cumprod(R)
-	return R
+    R = 1 - d/r
+    R[np.isnan(R)] = 0
+    R = np.cumprod(R)
+    return R
 
 def turnbull(x, c, n, t, estimator='Kaplan-Meier'):
-	
-	bounds = np.unique(np.concatenate([np.unique(x), np.unique(t)]))
-	N = n.sum()
+    bounds = np.unique(np.concatenate([np.unique(x), np.unique(t)]))
+    N = n.sum()
 
-	if x.ndim == 1:
-		x_new = np.empty(shape=(x.shape[0], 2))
-		x_new[:, 0] = x
-		x_new[:, 1] = x
-		x = x_new
+    if x.ndim == 1:
+        x_new = np.empty(shape=(x.shape[0], 2))
+        x_new[:, 0] = x
+        x_new[:, 1] = x
+        x = x_new
 
-	# Unpack x array
-	xl = x[:, 0]
-	xr = x[:, 1]
+    # Unpack x array
+    xl = x[:, 0]
+    xr = x[:, 1]
 
-	# Unpack t array
-	tl = t[:, 0]
-	tr = t[:, 1]
-	
-	# If there are left and right censored observations, convert them to interval censored observations
-	xl[c == -1] = -np.inf
-	xr[c ==  1] =  np.inf
-	
-	# bounds = bounds[0:-1]
-	M = bounds.size
-	N = xl.size
-	
-	alpha = np.zeros(shape=(N, M))
-	beta  = np.ones(shape=(N, M))
+    # Unpack t array
+    tl = t[:, 0]
+    tr = t[:, 1]
+    
+    # If there are left and right censored observations, convert them to interval censored observations
+    xl[c == -1] = -np.inf
+    xr[c ==  1] =  np.inf
+    
+    # bounds = bounds[0:-1]
+    M = bounds.size
+    N = xl.size
+    
+    alpha = np.zeros(shape=(N, M))
+    beta  = np.ones(shape=(N, M))
 
-	for i in range(0, N):
-		x1, x2 = xl[i], xr[i]
-		t1, t2 = tl[i], tr[i]
-		if x1 == x2:
-			alpha[i, :] = (bounds == x1).astype(int) * n[i]
-		elif x2 == np.inf:
-			alpha[i, :] = ((bounds > x1) & (bounds <= x2)).astype(int) * n[i]
-		else:
-			alpha[i, :] = ((bounds >= x1) & (bounds < x2)).astype(int) * n[i]
+    for i in range(0, N):
+        x1, x2 = xl[i], xr[i]
+        t1, t2 = tl[i], tr[i]
+        if x1 == x2:
+            alpha[i, :] = (bounds == x1).astype(int) * n[i]
+        elif x2 == np.inf:
+            alpha[i, :] = ((bounds > x1) & (bounds <= x2)).astype(int) * n[i]
+        else:
[…]
