### 2021-07-17 920be18
Slight change to surpyval model docs
