### 2021-07-18 9c93754
Semi-complete nonparametric docs section
