+++ docs/SurPyval Models.rst
@@ -6,23 +6,29 @@ To get started, let's import some useful packages, as such, for the rest of this
 
 .. code:: python
 
-	import surpyval as surv
-	import numpy as np
-	from matplotlib import pyplot as plt
+    import surpyval as surv
+    import numpy as np
+    from matplotlib import pyplot as plt
+
+
+Survival modelling with *surpyval* is very easy. This page will take you through a series of scenarios that can show you how to use the features of *surpyval* to get you the answers you need. The first example is if you simply have a list of event times and need to find the distribution of best fit.
 
 Non-Parametric
 --------------
 
-Complete Data
-^^^^^^^^^^^^^
+In each of the examples below, each of the ``KaplanMeier``, ``NelsonAalen``, or ``FlemingHarrington`` can be substituted with any of the others. It is the choice of the analyst which should be used. The 
+``Turnbull`` estimator has additional capabilities that can be used when you have right truncated, left censored, or interval censored data.
+
+Complete Data - Non-Parametric
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. code:: python
 
-	x = np.array([32, 33, 34, 35, 36, 37, 38, 39, 40, 42])
-	n = np.array([10, 33, 81, 161, 224, 289, 336, 369, 383, 389])
+    x = np.array([32, 33, 34, 35, 36, 37, 38, 39, 40, 42])
+    n = np.array([10, 33, 81, 161, 224, 289, 336, 369, 383, 389])
 
-	# Weibull's measurements are cumulative so we need to tranasform them
-	n = np.concatenate([[n[0]], np.diff(n)])
+    # Weibull's measurements are cumulative so we need to tranasform them
+    n = np.concatenate([[n[0]], np.diff(n)])
 
 Using this data we can esimtate the reliability, that is, the probability that a sample of steel will survive up to a given applied stress. So what does that mean?
 
@@ -34,18 +40,18 @@ For this we can use the Nelson-Aalen estimator of the hazard rate, then convert 
 
 .. code:: python 
 
-	bofors_steel_na = surv.NelsonAalen.fit(x, n=n)
+    bofors_steel_na = surv.NelsonAalen.fit(x, n=n)
 
-	plt.figure(figsize=(10, 7));
-	plt.ylabel('Survival Probability')
-	plt.xlabel('Stress [1.275kg/mm2]')
-	plt.ylim([0, 1])
-	plt.xlim([31, 42])
-	plt.step(bofors_steel_na.x, bofors_steel_na.R)
-	plt.title('Survival Prob vs Stress of Bofors Steel');
+    plt.figure(figsize=(10, 7));
+    plt.ylabel('Survival Probability')
+    plt.xlabel('Stress [1.275kg/mm2]')
+    plt.ylim([0, 1])
+    plt.xlim([31, 42])
+    plt.step(bofors_steel_na.x, bofors_steel_na.R)
+    plt.title('Survival Prob vs Stress of Bofors Steel');
 
 .. image:: images/nonp-1.png
-	:align: center
+    :align: center
 
 So what purpose is this?
 
@@ -53,13 +59,13 @@ With our non-parametric model of the Bofors steel. We can use this model to esti
 
 .. code:: python
 
-	print(str(bofors_steel_na.sf(34).round(4).item() * 100) + "%")
+    print(str(bofors_steel_na.sf(34).round(4).item() * 100) + "%")
 
 Which gives:
 
 .. code:: text
-	
-	80.15%
+    
+    80.15%
 
 The above shows that approximately 80% will survive up to a stress of 34. Therefore we will have an approximately 20% chance of our component failing in the design. 
 
[…]
