+++ surpyval/parametric/weibull.py
@@ -30,13 +30,18 @@ class Weibull_(ParametricFitter):
     def _parameter_initialiser(self, x, c=None, n=None, t=None, offset=False):
         log_x = np.log(x)
         log_x[np.isnan(log_x)] = -np.inf
+        if (2 in c) or (-1 in c):
+            heuristic = "Turnbull"
+        else:
+            heuristic = "Nelson-Aalen"
         if offset:
-            results =  mpp(dist=self, x=x, c=c, n=n, t=t, on_d_is_0=True, offset=True)
+            results = mpp(dist=self, x=x, c=c, n=n, t=t, 
+                          heuristic=heuristic, on_d_is_0=True, offset=True)
             return (results['gamma'], *results['params'])
         else:
             gumb = para.Gumbel.fit(log_x, c, n, t, how='MLE')
             if not gumb.res.success:
-                gumb = para.Gumbel.fit(log_x, c, n, t, how='MPP')
+                gumb = para.Gumbel.fit(log_x, c, n, t, how='MPP', heuristic=heuristic)
             mu, sigma = gumb.params
             alpha, beta = np.exp(mu), 1. / sigma
             if (np.isinf(alpha) | np.isnan(alpha)):
