+++ surpyval/parametric/gumbel.py
@@ -8,342 +8,346 @@ from surpyval import parametric as para
 from surpyval.parametric.parametric_fitter import ParametricFitter
 
 class Gumbel_(ParametricFitter):
-	def __init__(self, name):
-		self.name = name
-		self.k = 2
-		self.bounds = ((None, None), (0, None),)
-		self.support = (-np.inf, np.inf)
-		self.plot_x_scale = 'linear'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['mu', 'sigma']
-		self.param_map = {
-			'mu'    : 0,
-			'sigma' : 1
-		}
-
-	def _parameter_initialiser(self, x, c=None, n=None):
-		return para.Gumbel.fit(x, c, n, how='MPP').params
-
-	def sf(self, x, mu, sigma):
-		r"""
-
-		Surival (or Reliability) function for the Gumbel Distribution:
-
-		.. math::
-			R(x) = 1 - e^{e^{-\left ( x - \mu \right ) / \sigma}}
-
-		Parameters
-		----------
-
-		x : numpy array or scalar
-			The values of the random variables at which the survival function will be calculated 
-		mu : numpy array like or scalar
-			The location parameter of the distribution
-		sigma : numpy array like or scalar
-			The scale parameter of the distribution
-
-		Returns
-		-------
-
-		sf : scalar or numpy array 
-			The scalar value of the survival function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the survival function at each corresponding value in the input array.
-
-
-		Examples
-		--------
-		>>> import numpy as np
-		>>> from surpyval import Gumbel
-		>>> x = np.array([1, 2, 3, 4, 5])
-		>>> Gumbel.sf(x, 3, 2)
-		array([0.69220063, 0.54523921, 0.36787944, 0.19229565, 0.06598804])
-		"""
-		return np.exp(-np.exp((x - mu)/sigma))
-
-	def cs(self, x, X, mu, sigma):
-		return self.sf(x + X, mu, sigma) / self.sf(X, mu, sigma)
-
-	def ff(self, x, mu, sigma):
-		r"""
-
-		CDF (or Failure) function for the Gumbel Distribution:
-
-		.. math::
-			F(x) = e^{e^{-\left ( x - \mu \right )/\sigma}}
-
-		Parameters
-		----------
-
-		x : numpy array or scalar
-			The values of the random variables at which the survival function will be calculated 
-		mu : numpy array like or scalar
-			The location parameter of the distribution
-		sigma : numpy array like or scalar
-			The scale parameter of the distribution
-
-		Returns
-		-------
-
-		ff : scalar or numpy array 
-			The scalar value of the failure function of the distribution if a scalar was passed. If an array like object was passed then a numpy array is returned with the value of the failure function at each corresponding value in the input array.
-
-
-		Examples
-		--------
-		>>> import numpy as np
-		>>> from surpyval import Gumbel
-		>>> x = np.array([1, 2, 3, 4, 5])
-		>>> Gumbel.ff(x, 3, 2)
-		array([0.30779937, 0.45476079, 0.63212056, 0.80770435, 0.93401196])
-		"""
-		return 1 - np.exp(-np.exp((x - mu)/sigma))
-
-	def df(self, x, mu, sigma):
-		r"""
-
[…]
