### 2021-07-18 #3484 @derrynknife
Thanks again @MatthewReid854

- Typo fixed
- line 49 changed
- type fixed
- Fixed
- made to be two sentences.
- changed to True
