### 2021-07-18 #3484 @derrynknife
@MatthewReid854, would me sharing the jupyter notebook with all the documentation examples be helpful?
