### 2021-07-18 598a824
Changed paper to be in accordance with second review by Matt
