+++ paper.md
@@ -30,11 +30,11 @@ Survival analysis is being increasingly used by scientists, data scientists, eng
 
 # Statement of need
 
-*SurPyval* fills a gap in the Python ecosystem of survival analysis. Other survival analysis packages, e.g. *lifelines* [@davidson2019lifelines] and *reliability* [@matthew_reid_2021_5030847] offer excellent methods for most applications. SurPyval has the advantage of being able to handle arbitrary combinations of observed values, censoring and truncation. Further, *scipy* [@jones2001scipy] has yet to implement some basic features of survival analysis; concretely, it does not handle censored data. Therefore, there is a gap in the Python ecosystem for a package that is flexible to accomodate any arbitrary combination of observed failures (or deaths); left, right, or interval censored; and left or right truncated data with a single format. Another powerful feature of *SurPyval* is that it lets users select an estimation method for their circumstances. Maximum Likelihood Estimation (MLE) is used in most other applications, but *SurPyval* also implements Minimal Product Spacing, Method of Moments, Probability Plotting, Mean Square Error, and Expectation-Maximisation. This variety of estimation methods gives users of SurPyval greater flexibility in their choice of which parameter estimation technique to use when fitting distributions to their data. Commercial packages are well developed but can be expensive. R is excellent for survival analysis but many analysts now use Python as is explained in the *lifelines* paper. Therefore there is a need to have another flexible and open source python package to do survival analysis.
+*SurPyval* fills a gap in the Python ecosystem of survival analysis. Other survival analysis packages, e.g. *lifelines* [@davidson2019lifelines] and *reliability* [@matthew_reid_2021_5030847] offer excellent methods for most applications. SurPyval has the advantage of being able to handle arbitrary combinations of observed values, censoring and truncation. Further, *scipy* [@jones2001scipy] has yet to implement some basic features of survival analysis; concretely, it does not handle censored data. Therefore, there is a gap in the Python ecosystem for a package that is flexible to accomodate any arbitrary combination of observed failures (or deaths); left, right, or interval censored; and left or right truncated data with a single format. Another powerful feature of *SurPyval* is that it lets users select an estimation method for their circumstances. Maximum Likelihood Estimation (MLE) is used in most other applications, but *SurPyval* also implements Maximum Product of Spacing, Method of Moments, Probability Plotting, Mean Square Error, and Expectation-Maximisation. This variety of estimation methods gives users of SurPyval greater flexibility in their choice of which parameter estimation technique to use when fitting distributions to their data. Commercial packages are well developed but can be expensive. R is excellent for survival analysis but many analysts now use Python as is explained in the *lifelines* paper. Therefore there is a need to have another flexible and open source python package to do survival analysis.
 
 # Features
 
-*SurPyval* is grouped into two modules, these are parametric and non-parametric modules. For the parametric modeul *SurPyval* offers several methods to estimate parameters; these are Maximum Likelihood Estimation (MLE), Mean Square Error (MSE), Probability Plotting (MPP), Minimum Product Spacing (MPS), Method of Moments (MOM), and Expectation-Maximisation (EM). The EM is only used for mixture models.
+*SurPyval* is grouped into two modules, these are parametric and non-parametric modules. For the parametric module *SurPyval* offers several methods to estimate parameters; these are Maximum Likelihood Estimation (MLE), Mean Square Error (MSE), Method of Probability Plotting (MPP), Maximum Product of Spacing (MPS), Method of Moments (MOM), and Expectation-Maximisation (EM). The EM is only used for mixture models.
 
[…]
