### 2021-07-18 d39245b
Added section on using own CHF
