### 2021-07-18 75787fa
version 0.6.1. Bug fix in Distribution class
