### 2021-07-18 6802583
Added expoweibull to the tests
