+++ surpyval/tests/test_fit.py
@@ -1,9 +1,13 @@
 import pytest
 import numpy as np
-from surpyval import Gumbel, Normal, Weibull, LogNormal
-from surpyval import Logistic, LogLogistic, Uniform, Beta
+from surpyval import (
+    Gumbel, Normal, Weibull, LogNormal,
+    Logistic, LogLogistic, Uniform, 
+    Beta, ExpoWeibull)
 
-DISTS = [Gumbel, Normal, Weibull, LogNormal, Logistic, LogLogistic, Uniform, Beta]
+DISTS = [Gumbel, Normal, Weibull, LogNormal, 
+         Logistic, LogLogistic, Uniform, 
+         Beta, ExpoWeibull]
 parameter_sample_bounds = [((1, 20), (0.5, 5)),
                            ((1, 100), (0.5, 100)),
                            ((1, 100), (0.5, 20)),
@@ -11,7 +15,8 @@ parameter_sample_bounds = [((1, 20), (0.5, 5)),
                            ((1, 100), (0.5, 20)),
                            ((1, 100), (0.5, 20)),
                            ((1, 100), (1, 100)),
-                           ((.1, 30), (.1, 30)),]
+                           ((.1, 30), (.1, 30)),
+                           ((1, 30), (.1, 10), (0.5, 1.5)),]
 FIT_SIZES = [5000, 10000, 20000, 50000, 100000, 1000000]
 TOL = 0.2
 
