### 2021-07-19 d54d694
Finally impmlemented the general cb function
