+++ surpyval/parametric/weibull3p.py
@@ -1,188 +1,0 @@
-import autograd.numpy as np
-from scipy.stats import uniform
-from scipy.special import ndtri as z
-from scipy.special import gamma as gamma_func
-from scipy.optimize import minimize
-from scipy.stats import pearsonr
-import surpyval
-from surpyval import nonparametric as nonp
-from surpyval import parametric as para
-from surpyval.parametric.parametric_fitter import ParametricFitter
-from autograd import grad
-
-class Weibull3p_(ParametricFitter):
-	"""
-	class for the three parameter weibull distribution.
-	"""
-	def __init__(self, name):
-		self.name = name
-		self.k = 3
-		self.bounds = ((0, None), (0, None), (None, None),)
-		self.plot_x_scale = 'log'
-		self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-			0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-			0.9, 0.95, 0.99, 0.999, 0.9999]
-		self.param_names = ['alpha', 'beta', 'gamma']
-		self.param_map = {
-			'alpha' : 0,
-			'beta'  : 1,
-			'gamma' : 2
-		}
-
-	def _parameter_initialiser(self, x, c=None, n=None):
-		x, c, n = surpyval.xcn_handler(x, c, n)
-		diff = (np.max(x) - np.min(x))/10
-		init_mpp = Weibull3p.fit(x, c=c, n=n, how='MPP', heuristic='Turnbull').params
-		init = init_mpp[0], init_mpp[1], np.min(x) - diff
-		self.bounds = ((0, None), (0, None), (None, np.min(x)))
-		return init
-
-	def sf(self, x, alpha, beta, gamma):
-		return np.exp(-((x - gamma) / alpha)**beta)
-
-	def cs(self, x, X, alpha, beta, gamma):
-		return self.sf(x + X, alpha, beta, gamma) / self.sf(X, alpha, beta, gamma)
-
-	def ff(self, x, alpha, beta, gamma):
-		return 1 - np.exp(-((x - gamma) / alpha)**beta)
-
-	def df(self, x, alpha, beta, gamma):
-		return (beta / alpha) * ((x - gamma) / alpha)**(beta-1) \
-			* np.exp(-((x - gamma) / alpha)**beta)
-
-	def hf(self, x, alpha, beta, gamma):
-		return (beta / alpha) * ((x - gamma) / alpha)**(beta - 1)
-
-	def Hf(self, x, alpha, beta, gamma):
-		return ((x - gamma) / alpha)**beta
-
-	def qf(self, p, alpha, beta, gamma):
-		return alpha * (-np.log(1 - p))**(1/beta) + gamma
-
-	def mean(self, alpha, beta, gamma):
-		return alpha * gamma_func(1 + 1/beta) + gamma
-
-	def moment(self, n, alpha, beta, gamma):
-		# This is wrong.
-		return (alpha - gamma)**n * gamma_func(1 + n/beta)
-
-	def random(self, size, alpha, beta, gamma):
-		U = uniform.rvs(size=size)
-		return self.qf(U, alpha, beta, gamma)
-
-	def mpp_x_transform(self, x):
-		return np.log(x)
-
-	def mpp_y_transform(self, y):
-		return np.log(-np.log((1 - y)))
-
[…]
