+++ surpyval/parametric/parametric.py
@@ -9,6 +9,7 @@ from scipy.special import ndtri as z
 from surpyval import nonparametric as nonp
 
 import matplotlib.pyplot as plt
+from scipy.optimize import approx_fprime
 from matplotlib.ticker import FixedLocator
 
 def _round_vals(x):
@@ -59,7 +60,7 @@ class Parametric():
         """
         if type(x) == list:
             x = np.array(x)
-        return self.dist.sf(x, *self.params)
+        return self.dist.sf(x - self.gamma, *self.params)
 
     def ff(self, x):
         r"""
@@ -91,7 +92,7 @@ class Parametric():
         """
         if type(x) == list:
             x = np.array(x)
-        return self.dist.ff(x, *self.params)
+        return self.dist.ff(x - self.gamma, *self.params)
 
     def df(self, x): 
         r"""
@@ -123,7 +124,7 @@ class Parametric():
         """
         if type(x) == list:
             x = np.array(x)
-        return self.dist.df(x, *self.params)
+        return self.dist.df(x - self.gamma, *self.params)
 
     def hf(self, x):
         r"""
@@ -155,7 +156,7 @@ class Parametric():
         """
         if type(x) == list:
             x = np.array(x)
-        return self.dist.hf(x, *self.params)
+        return self.dist.hf(x - self.gamma, *self.params)
 
     def Hf(self, x):
         r"""
@@ -187,7 +188,7 @@ class Parametric():
         """
         if type(x) == list:
             x = np.array(x)
-        return self.dist.Hf(x, *self.params)
+        return self.dist.Hf(x - self.gamma, *self.params)
 
     def qf(self, p):
         r"""
@@ -216,7 +217,7 @@ class Parametric():
         """
         if type(p) == list:
             p = np.array(p)
-        return self.dist.qf(p, *self.params)
+        return self.dist.qf(p, *self.params) + self.gamma
 
     def cs(self, x, X):
         r"""
@@ -245,7 +246,7 @@ class Parametric():
         >>> model.cs(11, 10)
         0.00025840046151723767
         """
-        return self.dist.cs(x, X, *self.params)
+        return self.dist.cs(x - self.gamma, X - self.gamma, *self.params)
 
     def random(self, size):
         r"""
@@ -298,7 +299,7 @@ class Parametric():
         8.929795115692489
         """
         if not hasattr(self, '_mean'):
-            self._mean = self.dist.mean(*self.params)
+            self._mean = self.dist.mean(*self.params) + self.gamma
         return self._mean
 
     def moment(self, n):
@@ -367,18 +368,23 @@ class Parametric():
         """
         if self.method != 'MLE':
             raise Exception('Only MLE has confidence bounds')
+
+        hess_inv = np.copy(self.hess_inv)
         
-        pvars = self.hess_inv[np.triu_indices(self.hess_inv.shape[0])]
-        old_errstate = np.seterr(all='ignore')
+        pvars = hess_inv[np.triu_indices(hess_inv.shape[0])]
+        old_err_state = np.seterr(invalid='raise',
+                                  divide='raise',
+                                  over='raise',
+                                  under='raise')
+        
+        np.seterr(all='ignore')
 
         if hasattr(self.dist, 'R_cb'):
-            R_cb = lambda x: self.dist.R_cb(x, *self.params, self.hess_inv, alpha_ci=alpha_ci)
+            R_cb = lambda x: self.dist.R_cb(x - self.gamma, *self.params, hess_inv, alpha_ci=alpha_ci)
 
         else:
-
-
             def R_cb(x):
[…]

[…]
