### 2021-07-19 a017a39
Removed need for OffsetParametric class
