### 2021-07-19 c0ba572
Improved docs for user defined confidence bounds
