### 2021-07-19 #3484 @MatthewReid854
Article looks good now.

Question for @dfm , to what level of detail are we expected to scrutinize the code? I can run a few examples from the documentation and the output seems to be fine. As reviewers, how thorough are we expected to be?

Question for @derrynknife , how do I extract the offset parameter from the results as I have done below with alpha and beta?
```
model = surv.Weibull.fit(x, offset=True)
alpha = model.params[0]
beta = model.params[1]
```
