### 2021-07-19 #3484 @derrynknife
@MatthewReid854, it is stored as 'gamma':

`model = surv.Weibull.fit(x, offset=True)
alpha = model.params[0]
beta = model.params[1]`
