### 2021-07-20 #3484 @dfm
@MatthewReid854: Thanks for the question! Our main focus for this review is on the usage instructions and documentation. It's good to check the functionality claims (e.g. does the code run and give the expected results and performance), but you're definitely not expected to do a full code review. It sounds to me like you've done your due diligence, but let us know if there's anything else missing or unclear in the docs and examples. Thanks!!
