### 2021-07-21 #3484 @MatthewReid854
I am satisfied that the SurPyval python library functions as claimed in this paper. My testing is limited to the claims made in this paper and did not involve a thorough test of every function within the library.
I believe that the documentation is clear enough for a new user to become familiar with the API, though it could be improved with more examples and the inclusion of all required code in each code block. I found that many example code blocks are missing import statements if run in isolation.
Based on the limited testing I have done using examples in the paper and from the documentation, I accept the performance and functionality claims of this paper.
This concludes my review as I have ticked off all of the required checks.
