+++ surpyval/parametric/fitters/mle.py
@@ -1,7 +1,7 @@
 import autograd.numpy as np
-from autograd import jacobian, hessian
+from autograd import jacobian, hessian, value_and_grad
 from autograd.numpy.linalg import inv
-from scipy.optimize import minimize
+from scipy.optimize import minimize, approx_fprime
 
 import sys
 import copy
@@ -33,6 +33,8 @@ def mle(dist, x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset):
                               over='ignore',
                               under='ignore')
 
+    np.seterr(all='ignore')
+
     results = {}
 
     # Need to flag entries where truncation is inf or -inf so that the autograd doesn't fail. Because
@@ -51,10 +53,11 @@ def mle(dist, x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset):
     results['t_mle'] = t_mle
 
     # Create the objective function
-    def fun(params, offset=False, transform=True, gamma=0):
+    def fun(params, offset=False, transform=True, gamma=0.):
         x_mle = np.copy(x)
         if transform:
             params = inv_fs(const(params))
+
         if offset:
             gamma = params[0]
             params = params[1::]
@@ -77,23 +80,28 @@ def mle(dist, x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset):
             raise Exception
     except:
         # If first attempt fails, try a second time with only jacobian from autograd
-        # print("MLE with autodiff hessian and jacobian failed, trying without hessian", file=sys.stderr)
+        print("MLE with autodiff hessian and jacobian failed, trying without hessian", file=sys.stderr)
         try:
             res = minimize(fun, init, args=(offset, True), method='BFGS', jac=jac)
             if not res.success:
                 raise Exception
         except:
             # If second attempt fails, try a third time with only the objective function
-            # print("MLE with autodiff jacobian failed, trying without jacobian or hessian", file=sys.stderr)
+            print("MLE with autodiff jacobian failed, trying without jacobian or hessian", file=sys.stderr)
             try:
                 res = minimize(fun, init, args=(offset, True))
             except:
                 # Something really went wrong
-                # print("MLE FAILED: Likelihood function appears undefined; try alternate estimation method", file=sys.stderr)
+                print("MLE FAILED: Likelihood function appears undefined; try alternate estimation method", file=sys.stderr)
+                res = None
                 np.seterr(**old_err_state)
                 return {}
 
-    p_hat = inv_fs(const(res.x))
+    try:
+        p_hat = inv_fs(const(res.x))
+        print(p_hat)
+    except:
+        p_hat = [None] * len(init)
 
     if offset:
         results['gamma'] = p_hat[0]
@@ -107,7 +115,9 @@ def mle(dist, x, c, n, t, const, trans, inv_fs, init, fixed_idx, offset):
         results['hess_inv'] = inv(hessian(fun)(results['params'], *(False, False, results['gamma'])))
     except:
         results['hess_inv'] = None
+
     results['fun'] = fun
+    results['_neg_ll'] = res['fun']
     results['jac'] = jac
     results['res'] = res
 
