+++ surpyval/parametric/offset_parametric.py
@@ -1,548 +1,0 @@
-import re
-from autograd import jacobian
-import autograd.numpy as np
-from scipy.stats import uniform
-
-from surpyval import round_sig
-from scipy.special import ndtri as z
-from surpyval import nonparametric as nonp
-
-import matplotlib.pyplot as plt
-from matplotlib.ticker import FixedLocator
-
-def round_vals(x):
-	not_different = True
-	i = 1
-	while not_different:
-		x_ticks = np.array(round_sig(x, i))
-		not_different = (np.diff(x_ticks) == 0).any()
-		i += 1
-	return x_ticks
-
-class OffsetParametric():
-	def __repr__(self):
-		return 'Offset Parametric Surpyval model with {dist} distribution fitted by {method} yielding parameters {params} with offset of {gamma}'.format(dist=self.dist.name, method=self.method, params=self.params, gamma=self.gamma)
-
-	def sf(self, x):
-		return self.dist.sf(x - self.gamma, *self.params)
-
-	def ff(self, x):
-		return self.dist.ff(x - self.gamma, *self.params)
-
-	def df(self, x): 
-		return self.dist.df(x - self.gamma, *self.params)
-
-	def hf(self, x):
-		return self.dist.hf(x - self.gamma, *self.params)
-
-	def Hf(self, x):
-		return self.dist.Hf(x - self.gamma, *self.params)
-
-	def qf(self, p):
-		return self.dist.qf(p, *self.params) + self.gamma
-
-	def cs(self, x, X):
-		return self.dist.cs(x - self.gamma, X, *self.params)
-
-	def random(self, size):
-		return self.dist.qf(uniform.rvs(size=size), *self.params) + self.gamma
-
-	def mean(self):
-		return self.dist.mean(*self.params) + self.gamma
-
-	def moment(self):
-		return self.dist.moment(*self.params)
-
-	def entropy(self):
-		return self.dist.entropy(*self.params)
-
-	# def cb(self, x, sig):
-	# 	if self.method != 'MLE':
-	# 		raise Exception('Only MLE has confidence bounds')
-			
-	# 	du = z(sig) * np.sqrt(var_u)
-	# 	u_u = u_hat + du
-	# 	u_l = u_hat - du
-	# 	return np.vstack([np.exp(-np.exp(u_u)), np.exp(-np.exp(u_l))])
-
-	# def R_cb(self, x, cb=0.05):
-	# 	return self.dist.R_cb(x, *self.params, self.hess_inv, cb=cb)
-
-	def R_cb(self, t, cb=0.05):
-		"""
-		Need to add an H_cb function that can be used to convert to any
-		other function, i.e. hf, df, ff, sf OR....
-		Use this to convert. grad(log(R)) would work!
-		"""
-		sf_func = lambda params : self.dist.sf(t - params[0], *params[1::])
-
-		pvars = self.hess_inv[np.triu_indices(self.hess_inv.shape[0])]
-		with np.errstate(all='ignore'):
-			jac = jacobian(sf_func)(np.array([self.gamma, *self.params]))
-
-			
-		var_u = []
-		for i, j in enumerate(jac):
-			j = np.atleast_2d(j).T * j
-			j = j[np.triu_indices(j.shape[0])] 
-			var_u.append(np.sum(j * pvars))
-		diff = z(cb/2) * np.sqrt(np.array(var_u)) * np.array([1., -1.]).reshape(2, 1)
-		R_hat = self.sf(t)
[…]
