### 2021-07-25 992ebe3
Major stability improvement using log_Xf instead of functions
