+++ surpyval/parametric/parametric_fitter.py
@@ -24,1030 +24,546 @@ from .fitters import bounds_convert, fix_idx_and_function
 
 PARA_METHODS = ['MPP', 'MLE', 'MPS', 'MSE', 'MOM']
 
-class ParametricFitter0():
-	def like(self, x, c, n, *params):
-		like = np.zeros_like(x).astype(float)
-		like = np.where(c ==  0, self.df(x, *params), like)
-		like = np.where(c == -1, self.ff(x, *params), like)
-		like = np.where(c ==  1, self.sf(x, *params), like)
-		return like
-
-	def like_i(self, x, c, n, inf_c_flags, *params):
-		# This makes sure that any intervals that are at the boundaries of support or
-		# are infinite will not cause the autograd functions to fail.
-		ir = np.where(inf_c_flags[:, 1] == 1, 1, self.ff(x[:, 1], *params))
-		il = np.where(inf_c_flags[:, 0] == 1, 0, self.ff(x[:, 0], *params))
-		like_i = ir - il
-		return like_i
-
-	def like_t(self, t, t_flags, *params):
-		tr_denom = np.where(t_flags[:, 1] == 1, self.ff(t[:, 1], *params), 1.)
-		tl_denom = np.where(t_flags[:, 0] == 1, self.ff(t[:, 0], *params), 0.)
-		t_denom = tr_denom - tl_denom
-		return t_denom
-
-	def neg_ll(self, x, c, n, inf_c_flags, t, t_flags, *params):
-		print(params)
-		if 2 in c:
-			like_i = self.like_i(x, c, n, inf_c_flags, *params)
-			x_ = copy(x[:, 0])
-			x_[x_ == 0] = 1
-		else:
-			like_i = 0
-			x_ = copy(x)
-		
-		like = self.like(x_, c, n, *params)
-		like = like + like_i
-		like = np.log(like) - np.log(self.like_t(t, t_flags, *params))
-		like = np.multiply(n, like)
-		like = -np.sum(like)
-		return like
-
-	def neg_mean_D(self, x, c, n, *params):
-		mask = c == 0
-		x_obs = x[mask]
-		n_obs = n[mask]
-
-		gamma = 0
-
-		# Assumes already ordered
-		F  = self.ff(x_obs - gamma, *params)
-		D0 = F[0]
-		Dn = 1 - F[-1]
-		D = np.diff(F)
-		D = np.concatenate([np.array([D0]), D.T, np.array([Dn])]).T
-
-		Dr = self.sf(x[c ==  1]  - gamma, *params)
-		Dl = self.ff(x[c == -1]  - gamma, *params)
-
-		if (n_obs > 1).any():
-			n_ties = (n_obs - 1).sum()
-			Df = self.df(x_obs  - gamma, *params)
-			# Df = Df[Df != 0]
-			LL = np.concatenate([Dl, Df, Dr])
-			ll_n = np.concatenate([n[c == -1], (n_obs - 1), n[c == 1]])
-		else:
-			Df = []
-			n_ties = n_obs.sum()
-			LL = np.concatenate([Dl, Dr])
-			ll_n = np.concatenate([n[c == -1], n[c == 1]])
-		
-
-		M = np.log(D)
-		M = -np.sum(M)/(M.shape[0])
-		
-		LL = -(np.log(LL) * ll_n).sum()/(n.sum() - n_obs.sum() + n_ties)
-		return M + LL
-
-	def _moment(self, n, *params, offset=False):
-		if offset:
-			gamma = params[0]
-			params = params[1::]
-			fun = lambda x : x**n * self.df((x - gamma), *params)
-			m = quad(fun, gamma, np.inf)[0]
-		else:
-			if hasattr(self, 'moment'):
-				m = self.moment(n, *params)
[…]
