### 2021-07-26 1b4a32a
removed line that eliminated impact of 0 values on log_like. In prep for ds and lfp
