### 2021-07-26 93b95a3
Change main fit method to be much shorter and a little more readable
