### 2021-07-27 12499e8
Changed default plotting to turnbull-FH after learning how poorly NA does in ZI models. Completed ZI and LFP general model fit.
