+++ surpyval/parametric/parametric_fitter.py
@@ -48,19 +48,30 @@ class ParametricFitter():
         like = np.where(c ==  1, self.sf(x, *params), like)
         return like
 
-    def log_like(self, x, c, n, p, *params):
+    def log_like(self, x, c, n, p, f0, *params):
+        # This is getting a bit much....
         like = np.zeros_like(x).astype(float)
-        like = np.where(c ==  0, np.log(p) + self.log_df(x, *params), like)
-        like = np.where(c == -1, np.log(p) + self.log_ff(x, *params), like)
-        if p == 1:
-            like = np.where(c ==  1, self.log_sf(x, *params), like)
+        if f0 == 0:
+            like = np.where(c ==  0, np.log(p) + self.log_df(x, *params), like)
+            like = np.where(c == -1, np.log(p) + self.log_ff(x, *params), like)
+            if p == 1:
+                like = np.where(c ==  1, self.log_sf(x, *params), like)
+            else:
+                like = np.where(c ==  1, np.log(1 - (p * self.ff(x, *params))), like)
         else:
-            like = np.where(c ==  1, np.log(1 - (p * self.ff(x, *params))), like)
+            like = np.where(c ==  0, np.log(1 - f0) + np.log(p) + self.log_df(x, *params), like)
+            like = np.where(c == -1, np.log(1 - f0) + np.log(p) + self.log_ff(x, *params), like)
+            like = np.where((c == 0) & (x == 0), np.log(f0), like)
+            if p == 1:
+                like = np.where(c ==  1, np.log(1 - f0) + self.log_sf(x, *params), like)
+            else:
+                like = np.where(c ==  1, np.log(1 - f0) + np.log(1 - (p * self.ff(x, *params))), like)
+
         return like
 
-    def log_like_i(self, x, c, n, inf_c_flags, p, *params):
-        ir = np.where(inf_c_flags[:, 1] == 1, 1, p * self.ff(x[:, 1], *params))
-        il = np.where(inf_c_flags[:, 0] == 1, 0, p * self.ff(x[:, 0], *params))
+    def log_like_i(self, x, c, n, inf_c_flags, p, f0, *params):
+        ir = np.where(inf_c_flags[:, 1] == 1, 1, (1 - f0) * p * self.ff(x[:, 1], *params))
+        il = np.where(inf_c_flags[:, 0] == 1, 0, (1 - f0) * p * self.ff(x[:, 0], *params))
         like_i = ir - il
         like_i = np.where(c != 2, 1., like_i)
         return np.log(like_i)
@@ -74,12 +85,14 @@ class ParametricFitter():
         return like_i
 
     def like_t(self, t, t_flags, *params):
+        # Needs to be updated to work with zi and ds models.
+        # until then, can prevent it working in the `fit` method
         tr_denom = np.where(t_flags[:, 1] == 1, self.ff(t[:, 1], *params), 1.)
         tl_denom = np.where(t_flags[:, 0] == 1, self.ff(t[:, 0], *params), 0.)
         t_denom = tr_denom - tl_denom
         return t_denom
 
-    def neg_ll(self, x, c, n, inf_c_flags, t, t_flags, gamma, p, *params):
+    def neg_ll(self, x, c, n, inf_c_flags, t, t_flags, gamma, p, f0, *params):
         # if 2 in c:
         #     like_i = self.like_i(x, c, n, inf_c_flags, *params)
         #     x_ = copy(x[:, 0])
@@ -91,7 +104,7 @@ class ParametricFitter():
         x = copy(x) - gamma
 
         if 2 in c:
-            like_i = self.log_like_i(x, c, n, inf_c_flags, p, *params)
[…]
