+++ surpyval/parametric/parametric_fitter.py
@@ -186,7 +186,7 @@ class ParametricFitter():
             offset=False, zi=False, lfp=False, tl=None, tr=None,
             xl=None, xr=None, fixed=None, heuristic='Turnbull',
             init=[], rr='y', on_d_is_0=False,
-            turnbull_estimator='Nelson-Aalen'):
+            turnbull_estimator='Fleming-Harrington'):
 
         r"""
 
@@ -247,14 +247,20 @@ class ParametricFitter():
         init : array like, optional
             initial guess of parameters. Useful if method is failing.
 
-        rr : {'y', 'x'}
+        rr : ('y', 'x')
             The dimension on which to minimise the spacing between the line and the observation.
             If 'y' the mean square error between the line and vertical distance to each point is minimised.
             If 'x' the mean square error between the line and horizontal distance to each point is minimised.
 
-        turnbull_estimator : ('Nelson-Aalen' or 'Kaplan-Meier'), str, optional
-            If using the Turnbull heuristic, you can elect to use either the KM or NA estimator with 
-            the Turnbull estimate of x, r, and d. Defaults to KM.
+        on_d_is_0 : boolean, optional
+            For the case when using MPP and the highest value is right censored, you can choosed to
+            include this value into the regression analysis or not. That is, if :code:`False`, all values
+            where there are 0 deaths are excluded from the regression. If :code:`True` all values
+            regardless of whether there is a death or not are included in the regression.
+
+        turnbull_estimator : ('Nelson-Aalen', 'Kaplan-Meier', or 'Fleming-Harrington'), str, optional
+            If using the Turnbull heuristic, you can elect to use either the KM, NA, or FH estimator with 
+            the Turnbull estimates of r, and d. Defaults to FH.
 
         Returns
         -------
@@ -263,43 +269,50 @@ class ParametricFitter():
             A parametric model with the fitted parameters and methods for all functions of the distribution using the 
             fitted parameters.
 
-
         Examples
         --------
         >>> from surpyval import Weibull
+        >>> import numpy as np
         >>> x = Weibull.random(100, 10, 4)
         >>> model = Weibull.fit(x)
         >>> print(model)
-        Parametric Surpyval model with Weibull distribution fitted by MLE yielding parameters [10.25563233  3.68512055]
-
-        >>> model = Weibull.fit(x, how='MPS', fixed={'alpha' : 10})
-        >>> print(model)
-        Parametric Surpyval model with Weibull distribution fitted by MPS yielding parameters [10.          3.45512434]
-
-        >>> model = Weibull.fit(xl=x, xr=x+2, how='MPP')
-        >>> print(model)
-        Parametric Surpyval model with Weibull distribution fitted by MPP yielding parameters (11.465337182989183, 9.217130068358955)
-
+        Parametric SurPyval Model
+        =========================
+        Distribution        : Weibull
+        Fitted by           : MLE
+        Parameters          :
+             alpha: 10.551521182640098
+              beta: 3.792549834495306
+        >>> Weibull.fit(x, how='MPS', fixed={'alpha' : 10})
+        Parametric SurPyval Model
+        =========================
+        Distribution        : Weibull
+        Fitted by           : MPS
+        Parameters          :
+             alpha: 10.0
+              beta: 3.4314657446866836
+        >>> Weibull.fit(xl=x-1, xr=x+1, how='MPP')
+        Parametric SurPyval Model
+        =========================
+        Distribution        : Weibull
+        Fitted by           : MPP
+        Parameters          :
+             alpha: 9.943092756713078
+              beta: 8.613016934518258
         >>> c = np.zeros_like(x)
[…]
