+++ surpyval/nonparametric/nonparametric.py
@@ -5,358 +5,368 @@ from .kaplan_meier import KaplanMeier
 from .nelson_aalen import NelsonAalen
 from .fleming_harrington import FlemingHarrington_
 
+from scipy.interpolate import interp1d
+
 import matplotlib.pyplot as plt
 import pandas as pd
 
 class NonParametric():
-	"""
-	Result of ``.fit()`` method for every non-parametric surpyval distribution. This means that each of the 
-	methods in this class can be called with a model created from the ``NelsonAalen``, ``KaplanMeier``, 
-	``FlemingHarrington``, or ``Turnbull`` estimators.
-	"""
-	def __repr__(self):
-		return "{model} survival model".format(model=self.model)
-
-	def sf(self, x, how='step'):
-		r"""
-
-		Surival (or Reliability) function with the non-parametric estimates from the data
-
-		Parameters
-		----------
-
-		x : array like or scalar
-			The values of the random variables at which the survival function will be calculated 
-
-		Returns
-		-------
-
-		sf : scalar or numpy array 
-			The value(s) of the survival function at each x
-
-
-		Examples
-		--------
-		>>> from surpyval import NelsonAalen
-		>>> x = np.array([1, 2, 3, 4, 5])
-		>>> model = NelsonAalen.fit(x)
-		>>> model.sf(2)
-		array([0.63762815])
-		>>> model.sf([1., 1.5, 2., 2.5])
-		array([0.81873075, 0.81873075, 0.63762815, 0.63762815])
-		"""
-		x = np.atleast_1d(x)
-		# Let's not assume we can predict above the highest measurement
-		if how == 'step':
-			idx = np.searchsorted(self.x, x, side='right') - 1
-			R = self.R[idx]
-			R[np.where(x < self.x.min())] = 1
-			R[np.where(x > self.x.max())] = np.nan
-			R[np.where(x < 0)] = np.nan
-			return R
-		elif how == 'interp':
-			R = np.hstack([[1], self.R])
-			x_data = np.hstack([[0], self.x])
-			R = np.interp(x, x_data, R)
-			R[np.where(x > self.x.max())] = np.nan
-			return R
-
-	def ff(self, x, how='step'):
-		r"""
-
-		CDF (failure or unreliability) function with the non-parametric estimates from the data
-
-		Parameters
-		----------
-
-		x : array like or scalar
-			The values of the random variables at which the survival function will be calculated 
-
-		Returns
-		-------
-
-		ff : scalar or numpy array 
-			The value(s) of the failure function at each x
-
-
-		Examples
-		--------
-		>>> from surpyval import NelsonAalen
-		>>> x = np.array([1, 2, 3, 4, 5])
-		>>> model = NelsonAalen.fit(x)
-		>>> model.ff(2)
-		array([0.36237185])
-		>>> model.ff([1., 1.5, 2., 2.5])
-		array([0.18126925, 0.18126925, 0.36237185, 0.36237185])
-		"""
-		return 1 - self.sf(x, how=how)
-
-	def hf(self, x, how='step'):
-		r"""
-
-		Instantaneous hazard function with the non-parametric estimates from the data. This is
-		calculated using simply the difference between consecutive H(x).
-
-		Parameters
-		----------
-
-		x : array like or scalar
-			The values of the random variables at which the survival function will be calculated 
-
-		Returns
-		-------
-
-		hf : scalar or numpy array 
-			The value(s) of the failure function at each x
-
-
-		Examples
-		--------
-		>>> from surpyval import NelsonAalen
-		>>> x = np.array([1, 2, 3, 4, 5])
-		>>> model = NelsonAalen.fit(x)
[…]
