### 2021-07-27 8c14206
Added change log, and started application docs. Created check in user defined distribution
