### 2021-07-27 e8ab1eb
Added actuary and demography example to applications docs
