+++ docs/applications.rst
@@ -1,14 +1,97 @@
 Example Applications
 ====================
 
-The parametric API is essentially the exact same as the non-parametric API. All models are fit by a 
-call to the ``fit()`` method. However, the parametric models have more options that are only applicable to parametric modelling. The inputs of ``x`` for the random variable, ``c`` for the censoring flag, ``n``
-for count of each ``x``, ``xl`` and ``xr`` for intervally censored data (can't be used with ``x``) ``t``
-for the truncation matrix, ``tl`` for the left truncation scalar or array, and ``tr`` for the right truncation scalar or array all remain.
+This section documents some of the applications that a survival analysis toolkit can be useful to you, no matter what discipline you need it for.
 
 Reliability Engineering
 -----------------------
 
 Demographics / Actuarial
 ------------------------
+
+In demographics and actuarial studies, the distribution of the life of a population is of interest. For the demographer, it is necessary to understand how a population might change, in particular, how the expected lifespan is changing over time. The same applies to an actuary, an actuary is interested in lifetimes to understand the risk of payouts among those who own a life insurance policy.
+
+The `Gompertz-Makeham <https://en.wikipedia.org/wiki/Gompertz–Makeham_law_of_mortality>`_ is a distribution used in demography and actuarial studies to estimate the lifetime of a population. This can be implemented in surpyval with relative ease.
+
+.. code:: python
+
+
+    import surpyval as surv
+    from autograd import numpy as np
+    from matplotlib import pyplot as plt
+    from scipy.special import lambertw
+
+    bounds = ((0, None), (0, None), (0, None),)
+    support = (0, np.inf)
+    param_names = ['lambda', 'alpha', 'beta']
+    def Hf(x, *params):
+        Hf = params[0] * x + (params[1]/params[2])*(np.exp(params[2]*x))
+        return Hf
+        
+    GompertzMakeham = surv.Distribution('GompertzMakeham', Hf, param_names, bounds, support)
+
+We now have a GM distribution object that can be used to fit data. But we need some data:
+
+.. code:: python
+
+    # GM qf()
+    def qf(p, params):
+        lambda_ = params[0]
+        alpha = params[1]
+        beta = params[2]
+        return (alpha/(lambda_ * beta) - (1./lambda_)*np.log(1 - p)
+                - (1./beta)*lambertw((alpha*np.exp(alpha/lambda_)*(1 - p)**(-(beta/lambda_)))/(lambda_))).real
+
+
+    np.random.seed(1)
+    params = np.array([.68, 28.7e-3, 102.3])/1000
+    x = qf(np.random.uniform(0, 1, 100_000), params)
+    # Filter out some numeric overflows.
+    x = x[np.isfinite(x)]
+
+
+The parameters for the distribution come from Gavrilov et al. ([GGN]_), specifically the parameters for the lifespans of the 1974-1978 data. So in this case we have (simulated) data on the lifespans of 100,000 thousand people and we need to determine the GM parameters. This can be compared to the historic parameters to see if the age related mortality has changed or has remained roughly constant. To do so, all we need do with surpyval is to put the data to the ``fit()`` method.
+
+.. code:: python
+
+    model = GompertzMakeham.fit(x)
+    model.plot(alpha_ci=0.99, heuristic='Nelson-Aalen')
+    model
+
+.. code:: text
+
+    Parametric SurPyval Model
+    =========================
+    Distribution        : GompertzMakeham
+    Fitted by           : MLE
+    Parameters          :
+        lambda: 0.0007827108147066848
+         alpha: 2.1199267751549727e-05
+          beta: 0.10690878152126947
+
+.. image:: images/applications-demography-1.png
+    :align: center
+
+
[…]
