+++ docs/index.rst
@@ -6,9 +6,7 @@
 SurPyval - Survival Analysis in Python
 ======================================
 
-*surpyval* is an implementation of survival analysis in Python. The intent of this was to see if I could actually make it, and therefore learn a lot about survival statistics along the way, but also so that each time a model is created, it can be reused by other planned projects for monte carlo simulations (used in reliability engineering) and optimisations.
-
-Specifically, surpyval was designed to be used for the front end statistical analysis for the repyability package. The repyability package is a reliability engineering software package for engineers.
+*surpyval* is an implementation of survival analysis in Python. The intent of this was to see if I could actually make it, and therefore learn a lot about survival analysis along the way, but also so that each time a model is created, it can be reused by other planned projects for monte carlo simulations (used in reliability engineering) and optimisations.
 
 One feature of surpyval that separates it from other survival analysis packages is the intuitive way with which you can pass data to the fit methods. There are many different formats that can be used for survival analysis; surpyval handles many of the conceivable ways you can have your data stored. This is discussed in the data format tab.
 
@@ -102,7 +100,7 @@ Contents:
    
    Non-Parametric SurPyval Modelling
    Parametric SurPyval Modelling
-
+   applications
 
 .. toctree::
    :maxdepth: 1
