+++ docs/Contributing.rst
@@ -1,5 +1,5 @@
-Conntributing
-=============
+Contributing
+============
 
 If you want to contribute to SurPyval, please do! Please review the current open `feature reqeusts 
 <https://github.com/derrynknife/SurPyval/issues?q=is%3Aissue+is%3Aopen+label%3Aenhancement>`_ to see if your desired feature is in the requests. If not, please raise a new one to notify the community. We can assign you feature for you to branch and develop.
