### 2021-07-28 6790671
Doc updates, added aide-memoir page, updated changelog and applications
