+++ docs/applications.rst
@@ -6,6 +6,8 @@ This section documents some of the applications that a survival analysis toolkit
 Reliability Engineering
 -----------------------
 
+In 
+
 Demographics / Actuarial
 ------------------------
 
