### 2021-07-28 6a748a2
Attempted better initial guessing if using LFP
