### 2021-07-28 d04230c
Revamped nonparametric section and removed awkward duplicates of nonp estimators in turnbull.py
