+++ surpyval/nonparametric/kaplan_meier.py
@@ -1,14 +1,18 @@
-import surpyval
 import numpy as np
-from surpyval import nonparametric as nonp
+from surpyval.utils import xcnt_to_xrd
 from surpyval.nonparametric.nonparametric_fitter import NonParametricFitter
 
-def kaplan_meier(x, c=None, n=None, **kwargs):
-    # Could consider accepting on xrd for this
-    x, r, d = surpyval.xcnt_to_xrd(x, c, n, **kwargs)
-    
-    R = np.cumprod(1 - d/r)
-    return x, r, d, R
+def km(r, d):
+    R = 1 - d/r
+    R[np.isnan(R)] = 0
+    R = np.cumprod(R)
+    return R
+
+def kaplan_meier(x, c, n, t, **kwargs):
+    xrd = xcnt_to_xrd(x, c, n, t)
+    out = {k : v for k, v in zip(['x', 'r', 'd'], xrd)}
+    out['R'] = km(out['r'], out['d'])
+    return out
 
 class KaplanMeier_(NonParametricFitter):
     r"""
