### 2021-07-28 dee04fb
Updated __repr__ function for NonParametric
