### 2021-07-28 55cdd3f
Added docstring to NonParametric fit() method
