+++ surpyval/nonparametric/nonparametric_fitter.py
@@ -3,10 +3,77 @@ from surpyval import nonparametric as nonp
 from surpyval.utils import xcnt_handler
 
 class NonParametricFitter():
-    
     def fit(self, x=None, c=None, n=None, t=None,
             xl=None, xr=None, tl=None, tr=None,
-            estimator='Fleming-Harrington'):
+            turnbull_estimator='Fleming-Harrington'):
+        r"""
+
+        The central feature to SurPyval's capability. This function aimed to have an API to mimic the 
+        simplicity of the scipy API. That is, to use a simple :code:`fit()` call, with as many or as few
+        parameters as is needed.
+
+        Parameters
+        ----------
+
+        x : array like, optional
+            Array of observations of the random variables. If x is :code:`None`, xl and xr must be provided.
+
+        c : array like, optional
+            Array of censoring flag. -1 is left censored, 0 is observed, 1 is right censored, and 2 is intervally
+            censored. If not provided will assume all values are observed.
+
+        n : array like, optional
+            Array of counts for each x. If data is proivded as counts, then this can be provided. If :code:`None`
+            will assume each observation is 1.
+
+        t : 2D-array like, optional
+            2D array like of the left and right values at which the respective observation was truncated. If
+            not provided it assumes that no truncation occurs.
+
+        tl : array like or scalar, optional
+            Values of left truncation for observations. If it is a scalar value assumes each observation is
+            left truncated at the value. If an array, it is the respective 'late entry' of the observation
+
+        tr : array like or scalar, optional
+            Values of right truncation for observations. If it is a scalar value assumes each observation is
+            right truncated at the value. If an array, it is the respective right truncation value for each
+            observation
+
+        xl : array like, optional
+            Array like of the left array for 2-dimensional input of x. This is useful for data that is all
+            intervally censored. Must be used with the :code:`xr` input.
+
+        xr : array like, optional
+            Array like of the right array for 2-dimensional input of x. This is useful for data that is all
+            intervally censored. Must be used with the :code:`xl` input.
+
+        turnbull_estimator : ('Nelson-Aalen', 'Kaplan-Meier', or 'Fleming-Harrington'), str, optional
+            If using the Turnbull heuristic, you can elect to use either the KM, NA, or FH estimator with 
+            the Turnbull estimates of r, and d. Defaults to FH.
+
+        Returns
+        -------
+
+        model : NonParametric
+            A parametric model with the fitted parameters and methods for all functions of the distribution using the 
+            fitted parameters.
+
+        Examples
+        --------
+        >>> from surpyval import NelsonAalen, Weibull, Turnbull
+        >>> import numpy as np
+        >>> x = Weibull.random(100, 10, 4)
+        >>> model = NelsonAalen.fit(x)
+        >>> print(model)
+        Non-Parametric SurPyval Model
+        =============================
+        Model            : Nelson-Aalen
+        >>> Turnbull.fit(x, turnbull_estimator='Kaplan-Meier')
+        Non-Parametric SurPyval Model
+        =============================
+        Model            : Turnbull
+        Estimator        : Kaplan-Meier
+        """
         x, c, n, t = xcnt_handler(x=x, c=c, n=n, t=t, 
                                   tl=tl, tr=tr, 
                                   xl=xl, xr=xr)
@@ -17,7 +84,7 @@ class NonParametricFitter():
         data['n'] = n
         data['t'] = t
         if self.how == 'Turnbull':
-            data['estimator'] = estimator
+            data['estimator'] = turnbull_estimator
 
         out = nonp.NonParametric()
         out.data = data
