### 2021-07-28 92871dd
added fsli_to_xcn function
