### 2021-07-28 2fb0a3d
Created cool reliability engineering example in the applications docs
