### 2021-07-28 b2edd84
Fixed bug where if fit fails it cannot assign parameters to the model
