+++ surpyval/parametric/parametric_fitter.py
@@ -441,7 +441,7 @@ class ParametricFitter():
         for k, v in results.items():
             setattr(model, k, v)
 
-        if hasattr(mdoel, 'params'):
+        if hasattr(model, 'params'):
             for k, v in zip(self.param_names, model.params):
                 setattr(model, k, v)
 
