### 2021-07-28 d17d334
Bug on right truncation plotting fixed
