+++ docs/applications.rst
@@ -1,14 +1,170 @@
 Example Applications
 ====================
 
-This section documents some of the applications that a survival analysis toolkit can be useful to you, no matter what discipline you need it for.
+This section documents some of the applications that SurPyval as a survival analysis toolkit can be useful to you, no matter what discipline you need it for.
 
-Reliability Engineering
------------------------
+Boston House Prices
+-------------------
 
-In reliability engineering we might be interested in the proportion of a population that will experience a particular failure mode. Further, in an effort to save money, we will want to determine the minimum duration of a test that can establish whether a component will fail. If the test is too long we will waste time and money in testing and if the testing is too short we will ship some products that will fail in the field. We need to optimise this interval.
+No statistical analysis package can avoid doing the 'hello world' task of analysing the 'Boston House Pricing' dataset. What might surprise some readers is that this would even be considered... The myriad blogs and Kaggle posts looking into this problem can not surely be improved upon. I agree, however, it is a good example of why one needs to be aware of censoring and how flexible the SurPyval API is when dealing with it. Looking at the boston house pricing dataset you can see that there is a suspicous number of houses at the top end that all have the same price:
 
-Using data from the paper that introduced the Limited Failure Population model to the reliability engineering world [Meeker]_ we can show how surpyval can be used to calculate the test duration.
+.. image:: images/applications-boston-1.png
+    :align: center
+
+On consideration, one can see that there are no houses above $50,000 and that the density at that point is much higher than we would expect because we would expect some form of a 'fat-tail.' That is, we should expect a decreasing number of houses at the highest costs. It is therefore safe to conclude that all values above $50,000 have been set to $50,000; which is to say that the sale price is right censored! And because it is a censored observation we will need to use an analysis tool that can handle censored observations, lest we may be wrong in our estimates of the distribution of housing prices. So lets load the data and see what results we can get, starting with the raw data.
+
+
+.. code:: python
+
+    import surpyval as surv
+    x = Boston.df['medv'].values
+    x, c, n = surv.xcn_handler(x)
+
+    model = surv.Weibull.fit(x, c, n)
+    print(model)
+    model.plot()
+
+
+.. code:: text
+
+    Parametric SurPyval Model
+    =========================
+    Distribution        : Weibull
+    Fitted by           : MLE
+    Parameters          :
+         alpha: 25.386952832397032
+          beta: 2.5651903209947684
+
+
+.. image:: images/applications-boston-2.png
+    :align: center
+
+
+From the above plot you can see that near 50, the parametric model diverges substantially from the actual data. So we can see that having not censored the highest values means that our model could be improved by doing so. Let's see:
+
+
+.. code:: python
+
+    import surpyval as surv
+    x = Boston.df['medv'].values
+    x, c, n = surv.xcn_handler(x)
+    # Right censor the highest value
+    c[-1] = 1
+
+    model = surv.Weibull.fit(x, c, n)
+    print(model)
+    model.plot()
+
+
+.. code:: text
+
+    Parametric SurPyval Model
+    =========================
+    Distribution        : Weibull
+    Fitted by           : MLE
+    Parameters          :
+         alpha: 25.53669601993307
+          beta: 2.469159446459548
+
+
+.. image:: images/applications-boston-3.png
+    :align: center
+
+We can see that the model has changed slightly, however, there appears to be a 'disconnect' near 24 that makes the model a poor fit above and below the value. Let's see if a different distribution will improve the fit:
+
+
+.. code:: python
+
+    import surpyval as surv
[…]
