### 2021-07-29 ac0ddd6
fixed bug in R_cb of Gumbel
