+++ surpyval/parametric/gumbel.py
@@ -341,13 +341,13 @@ class Gumbel_(ParametricFitter):
             2 * z_hat * cv_matrix[0, 1])
         return var_z
 
-    def z_cb(self, x, mu, sigma, cv_matrix, cb=0.05):
+    def z_cb(self, x, mu, sigma, cv_matrix, alpha_ci=0.05):
         z_hat = (x - mu)/sigma
         var_z = self.var_z(x, mu, sigma, cv_matrix)
-        bounds = z_hat + np.array([1., -1.]).reshape(2, 1) * z(cb/2) * np.sqrt(var_z)
+        bounds = z_hat + np.array([1., -1.]).reshape(2, 1) * z(alpha_ci/2) * np.sqrt(var_z)
         return bounds
 
     def R_cb(self, x, mu, sigma, cv_matrix, alpha_ci=0.05):
-        return self.sf(self.z_cb(x, mu, sigma, cv_matrix, alpha_ci=0.05), 0, 1).T
+        return self.sf(self.z_cb(x, mu, sigma, cv_matrix, alpha_ci=alpha_ci), 0, 1).T
 
 Gumbel = Gumbel_('Gumbel')
