### 2021-07-29 7d24c5f
Added second reliability engineering example to applications docs
