### 2021-07-29 78aba53
Added criminology section to docs
