### 2021-07-29 19c0d81
Added economics example
