### 2021-07-31 a246a39
Fixed bug in NonParametric that required the x input to be in order
