### 2021-07-31 6d12338
Fixed error in dist_from_hazard for inv_y_transform
