### 2021-07-31 1a7f7d8
Simplified bonds_convert and fixing functions
