### 2021-07-31 ec63edb
fixed bug in R_cb for Exponential
