### 2021-08-01 4feecb2
added parameter to xcnt_handler to not group and sort as this is needed for regression models
