### 2021-08-02 7cae416
Created accelerated life models for all distributions. wow
