+++ surpyval/regression/accelerated_life.py
@@ -1,0 +1,252 @@
+import autograd.numpy as np
+from autograd import jacobian, hessian, elementwise_grad
+from scipy.optimize import minimize
+import surpyval
+import inspect
+
+from ..parametric.fitters import bounds_convert, fix_idx_and_function
+from .regression import Regression
+
+class ParameterSubstitutionFitter():
+    def __init__(self, kind, name, distribution, life_model, life_parameter, baseline=[]):
+
+        if str(inspect.signature(life_model.phi)) != '(X, *params)':
+            raise ValueError('PH function must have the signature \'(X, *params)\'')
+
+        if type(baseline) != list:
+            # Baseline used if using a function that deviates from some number, e.g. np.exp(np.dot(X, beta))
+            baseline = [baseline]
+
+        self.name = name
+        self.kind = kind
+        self.dist = distribution
+        self.life_model = life_model
+        self.k_dist = len(self.dist.param_names)
+        self.bounds = self.dist.bounds
+        self.support = self.dist.support
+        self.param_names = self.dist.param_names
+        self.param_map = {v : i for i, v in enumerate(self.dist.param_names)}
+        self.phi = life_model.phi
+        self.Hf_dist = self.dist.Hf
+        self.hf_dist = self.dist.hf
+        self.sf_dist = self.dist.sf
+        self.ff_dist = self.dist.ff
+        self.df_dist = self.dist.df
+        self.baseline = baseline
+        self.life_parameter = life_parameter
+        self.fixed = {life_parameter : 1.}
+
+    def Hf(self, x, X, *params):
+        x = np.array(x)
+        if np.isscalar(X):
+            X = np.ones_like(x) * X
+        else:
+            X = np.array(X)
+
+        dist_params = np.array(params[0:self.k_dist])
+        phi_params = np.array(params[self.k_dist:])
+
+        Hf = np.zeros_like(x)
+        stresses = np.unique(X, axis=0)
+        for stress in stresses:
+            life_param_mask = np.array(range(0, len(dist_params))) == self.param_map[self.life_parameter]
+            dist_params_i = np.where(life_param_mask,
+                self.phi(stress, *phi_params),
+                dist_params)
+            mask = (X == stress).all(axis=1)
+            Hf = np.where(mask, self.Hf_dist(x, *dist_params_i), Hf)
+
+        return Hf
+
+    def hf(self, x, X, *params):
+        x = np.array(x)
+        if np.isscalar(X):
+            X = np.ones_like(x) * X
+        else:
+            X = np.array(X)
+
+        dist_params = np.array(params[0:self.k_dist])
+        phi_params = np.array(params[self.k_dist:])
+
+        hf = np.zeros_like(x)
+        for stress in np.unique(X, axis=0):
+            life_param_mask = np.array(range(0, len(dist_params))) == self.param_map[self.life_parameter]
+            params = np.where(life_param_mask,
+                self.phi(stress, *phi_params),
+                dist_params)
+            mask = (X == stress).all(axis=1)
+            hf = np.where(mask, self.hf_dist(x, *params), hf)
+
+        return hf
+
+    def df(self, x, X, *params):
+        x = np.array(x)
+        if np.isscalar(X):
+            X = np.ones_like(x) * X
+        else:
+            X = np.array(X)
+        return self.hf(x, X, *params) * np.exp(-self.Hf(x, X, *params))
+
+    def sf(self, x, X, *params):
+        x = np.array(x)
+        if np.isscalar(X):
+            X = np.ones_like(x) * X
+        else:
+            X = np.array(X)
+        return np.exp(-self.Hf(x, X, *params))
+
+    def ff(self, x, X, *params):
+        x = np.array(x)
+        if np.isscalar(X):
[…]
