### 2021-08-02 9875308
Added docs update to aide memoir to explain the difference between al, aft, and ph models
