+++ surpyval/regression/accelerated_failure_time.py
@@ -1,0 +1,174 @@
+import autograd.numpy as np
+from autograd import jacobian, hessian, elementwise_grad
+from scipy.optimize import minimize
+import surpyval
+import inspect
+
+from ..parametric.fitters import bounds_convert, fix_idx_and_function
+from .regression import Regression
+
+class AcceleratedFailureTimeFitter():
+    def __init__(self, name, distribution, acc_model, fixed_parameter, 
+                 acc_parameter_relationship):
+
+        if str(inspect.signature(acc_model.phi)) != '(X, *params)':
+            raise ValueError('PH function must have the signature \'(X, *params)\'')
+
+        self.name = name
+        self.dist = distribution
+        self.acc_model = acc_model
+        self.acc_parameter_relationship = acc_parameter_relationship
+        self.fixed_parameter = fixed_parameter
+        self.k_dist = len(self.dist.param_names)
+        self.bounds = self.dist.bounds
+        self.support = self.dist.support
+        self.param_names = self.dist.param_names
+        self.param_map = {v : i for i, v in enumerate(self.dist.param_names)}
+        self.phi = acc_model.phi
+        self.Hf_dist = self.dist.Hf
+        self.hf_dist = lambda x, *params: elementwise_grad(self.Hf_dist)(x, *params)
+        self.sf_dist = lambda x, *params: np.exp(-self.Hf_dist(x, *params))
+        self.ff_dist = lambda x, *params: 1 - np.exp(-self.Hf_dist(x, *params))
+        self.df_dist = lambda x, *params: elementwise_grad(self.ff_dist)(x, *params)
+        self.fixed = {fixed_parameter : 1.}
+
+    def Hf(self, x, X, *params):
+        dist_params = np.array(params[0:self.k_dist])
+        phi_params = np.array(params[self.k_dist:])
+        return self.Hf_dist(self.phi(X, *phi_params) * x, *dist_params)
+
+    def hf(self, x, X, *params):
+        dist_params = np.array(params[0:self.k_dist])
+        phi_params = np.array(params[self.k_dist:])
+        return self.hf_dist(self.phi(X, *phi_params) * x, *dist_params)
+
+    def df(self, x, X, *params):
+        return self.hf(x, X, *params) * np.exp(-self.Hf(x, X, *params))
+
+    def sf(self, x, X, *params):
+        return np.exp(-self.Hf(x, X, *params))
+
+    def ff(self, x, X, *params):
+        return 1 - np.exp(-self.Hf(x, X, *params))
+
+    def _parameter_initialiser_dist(self, x, c=None, n=None, t=None):
+        out = []
+        for low, high in self.bounds:
+            if (low is None) & (high is None):
+                out.append(0)
+            elif high is None:
+                out.append(low + 1.)
+            elif low is None:
+                out.append(high - 1.)
+            else:
+                out.append((high + low)/2.)
+
+        return out
+
+    def mpp_inv_y_transform(self, y, *params):
+        return y
+
+    def mpp_y_transform(self, y, *params):
+        return y
+
+    def mpp_inv_y_transform(self, y, *params):
+        return y
+
+    def mpp_x_transform(self, x, gamma=0):
+        return x - gamma
+
+    def log_df(self, x, X, *params):
+        return np.log(self.hf(x, X, *params)) - self.Hf(x, X, *params)
+
+    def log_sf(self, x, X, *params):
+        return -self.Hf(x, X, *params)
+
+    def log_ff(self, x, X, *params):
+        return np.log(self.ff(x, X, *params))
+
+    def neg_ll(self, X, x, c, n, *params):
+        params = np.array(params)
+        like = np.zeros_like(x).astype(float)
[…]
