### 2021-08-04 f276e87
Fixed bug in way confidence bounds were calculated for offset distributions. It was not adjusting for the offset... ooops. Now fixed
