+++ docs/regression.rst
@@ -1,0 +1,10 @@
+
+Regression Analysis
+===================
+
+Regression analysis is the process of capturing the effect that other factors, i.e. covariates, will have on the survival probability. That is, we use data on other factors to 'regress' onto the survival distribution.
+
+SurPyval 
+
+Accelerated Life
+----------------
