### 2021-08-04 48317ec
Implemented general cb() method for NonParametric
