### 2021-08-04 5ac8f38
Improved docstrings for cb() methods in parametric and nonparametric
