### 2021-08-04 8c1308f
Added regression capability with AFT, AL, and PH fitters
