+++ surpyval/regression/__init__.py
@@ -1,4 +1,6 @@
 import autograd.numpy as np
+from collections import defaultdict
+
 from .cox_ph import CoxProportionalHazardsFitter
 
 from .proportional_hazards import ProportionalHazardsFitter
@@ -13,6 +15,7 @@ from .lifemodels.eyring import Eyring, InverseEyring
 from .lifemodels.dual_exponential import DualExponential
 from .lifemodels.dual_power import DualPower
 from .lifemodels.power_exponential import PowerExponential
+from .lifemodels.general_log_linear import GeneralLogLinear
 
 from ..parametric import (
     LogNormal,
@@ -25,10 +28,8 @@ from ..parametric import (
 # Semi-Parametric Proportional Hazard
 CoxPH = CoxProportionalHazardsFitter()
 
-DISTS = [surv.Exponential, LogNormal, Normal, 
-    Weibull, Gumbel, Logistic]
-LIFE_PARAMS = ['lambda', 'mu', 'mu', 'alpha', 'mu', 'mu']
-
+DISTS = [surv.Exponential, Normal, Weibull, Gumbel, Logistic, LogNormal]
+LIFE_PARAMS = ['lambda', 'mu', 'alpha', 'mu', 'mu', 'mu']
 LIFE_MODELS = [
     Power, 
     InversePower,
@@ -42,6 +43,16 @@ LIFE_MODELS = [
     PowerExponential
 ]
 
+life_parameter_transform = defaultdict(lambda: None)
+life_parameter_inverse_transform = defaultdict(lambda: None)
+baseline_parameters = defaultdict(lambda: [])
+
+life_parameter_transform['LogNormal'] = lambda x: np.log(x)
+life_parameter_transform['Exponential'] = lambda x: 1./x
+
+life_parameter_inverse_transform['LogNormal'] = lambda x: np.exp(x)
+life_parameter_inverse_transform['Exponential'] = lambda x: 1./x
+
 # Quite remarkable - creates every life model and distribution class!
 for dist, parameter in zip(DISTS, LIFE_PARAMS):
     for life_model in LIFE_MODELS:
@@ -49,43 +60,39 @@ for dist, parameter in zip(DISTS, LIFE_PARAMS):
         vars()[name] = ParameterSubstitutionFitter(
                             'Accelerated Life', 
                             name, dist, life_model, 
-                            parameter
+                            parameter,
+                            param_transform=life_parameter_transform[dist.name],
+                            inverse_param_transform=life_parameter_inverse_transform[dist.name]
                         )
 
-# Parametric Proportional Hazard
-def phi_param_map(X):
-    base = 'beta_'
-    return {base + str(i) : i for i in range(X.shape[1])}
+for dist, parameter in zip(DISTS, LIFE_PARAMS):
+    name = dist.name + life_model.name + "AL"
+    vars()[name] = ParameterSubstitutionFitter(
+                        'Accelerated Life', 
+                        name,
+                        dist,
+                        GeneralLogLinear,
+                        parameter,
+                        baseline=[parameter],
+                        param_transform=life_parameter_transform[dist.name],
+                        inverse_param_transform=life_parameter_inverse_transform[dist.name]
+                    )
 
+# Parametric Proportional Hazard
 ExponentialPH = ProportionalHazardsFitter(
     'ExponentialPH', 
     surv.Exponential,
     lambda X, *params: np.exp(np.dot(X, np.array(params))),
     lambda X: (((None, None),) * X.shape[1]),
-    phi_param_map=phi_param_map,
[…]
