### 2021-08-04 b584b8f
Fixed bug in MSE fitter created by changes to NonParametric
