### 2021-08-05 932ae24
Enhanced nonparametric to flake8 standards
