+++ surpyval/nonparametric/turnbull.py
@@ -1,9 +1,9 @@
 import numpy as np
 from surpyval.nonparametric.nonparametric_fitter import NonParametricFitter
-
 from .nelson_aalen import na
 from .kaplan_meier import km
 from .fleming_harrington import fh
+
 
 def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
     bounds = np.unique(np.concatenate([np.unique(x), np.unique(t)]))
@@ -22,17 +22,18 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
     # Unpack t array
     tl = t[:, 0]
     tr = t[:, 1]
-    
-    # If there are left and right censored observations, convert them to interval censored observations
+
+    # If there are left and right censored observations,
+    # convert them to interval censored observations
     xl[c == -1] = -np.inf
-    xr[c ==  1] =  np.inf
-    
+    xr[c == 1] = np.inf
+
     # bounds = bounds[0:-1]
     M = bounds.size
     N = xl.size
-    
+
     alpha = np.zeros(shape=(N, M))
-    beta  = np.ones(shape=(N, M))
+    beta = np.ones(shape=(N, M))
 
     for i in range(0, N):
         x1, x2 = xl[i], xr[i]
@@ -44,17 +45,16 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
         else:
             alpha[i, :] = ((bounds >= x1) & (bounds < x2)).astype(int) * n[i]
 
-        beta[i, :]  = (((bounds >= t1) & (bounds < t2)).astype(int))
+        beta[i, :] = (((bounds >= t1) & (bounds < t2)).astype(int))
 
-
-    beta[:, M-1] = 1
+    beta[:, M - 1] = 1
     n = n.reshape(-1, 1)
     d = np.zeros(M)
-    p = np.ones(M)/M
-    
+    p = np.ones(M) / M
+
     iters = 0
     p_prev = np.zeros_like(p)
-        
+
     if estimator == 'Kaplan-Meier':
         func = km
     elif estimator == 'Nelson-Aalen':
@@ -66,20 +66,21 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
 
     while (iters < 1000) & (not np.allclose(p, p_prev)):
         p_prev = p
-        iters +=1
+        iters += 1
         ap = alpha * p
         # Observed deaths
         mu = alpha * ap / ap.sum(axis=1, keepdims=True)
         # Expected additional failures due to truncation
-        nu = n*(1 - beta) * (1 - beta)*p / (beta*p).sum(axis=1, keepdims=True)
-        
+        nu = (n * (1 - beta) * (1 - beta) * p
+              / (beta * p).sum(axis=1, keepdims=True))
+
         # Deaths/Failures
         d = (nu + mu).sum(axis=0)
         # M total observed and unobserved failures.
         M = (nu + mu).sum()
         # Risk set, i.e the number of items at risk at each value x
         r = M - d.cumsum() + d
-        # Find the reliability using the deaths and risk set. Can use either NA or KM!
+        # Find the reliability using the deaths and risk set.
         R = func(r, d)
         # Calculate the probability mass in each interval
         p = np.abs(np.diff(np.hstack([[1], R])))
@@ -103,12 +104,13 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
 
     return out
 
+
 class Turnbull_(NonParametricFitter):
     r"""
-    Turnbull estimator class. Returns a `NonParametric` object from method :code:`fit()`.
-    Calculates the Non-Parametric estimate of the survival function using the Turnbull
-    NPMLE
-    
[…]
