+++ surpyval/nonparametric/nelson_aalen.py
@@ -2,25 +2,30 @@ import numpy as np
 from surpyval.utils import xcnt_to_xrd
 from surpyval.nonparametric.nonparametric_fitter import NonParametricFitter
 
+
 def na(r, d):
-    H = np.cumsum(d/r)
+    H = np.cumsum(d / r)
     H[np.isnan(H)] = np.inf
     R = np.exp(-H)
     return R
 
+
 def nelson_aalen(x, c, n, t):
     xrd = xcnt_to_xrd(x, c, n, t)
-    out = {k : v for k, v in zip(['x', 'r', 'd'], xrd)}
+    out = {k: v for k, v in zip(['x', 'r', 'd'], xrd)}
     out['R'] = na(out['r'], out['d'])
     return out
 
+
 class NelsonAalen_(NonParametricFitter):
     r"""
-    Nelson-Aalen estimator class. Returns a `NonParametric` object from method :code:`fit()` Calculates the Non-Parametric estimate of the survival function using:
+    Nelson-Aalen estimator class. Returns a `NonParametric`
+    object from method :code:`fit()` Calculates the Non-Parametric
+    estimate of the survival function using:
 
     .. math::
         R(x) = e^{-\sum_{i:x_{i} \leq x}^{} \frac{d_{i} }{r_{i}}}
-    
+
     Examples
     --------
     >>> import numpy as np
@@ -32,5 +37,6 @@ class NelsonAalen_(NonParametricFitter):
     """
     def __init__(self):
         self.how = 'Nelson-Aalen'
+
 
 NelsonAalen = NelsonAalen_()
