+++ surpyval/parametric/__init__.py
@@ -10,14 +10,11 @@ Parametric Analysis
 	model = para.Weibull.fit(x)
 
 """
-import autograd.numpy as np
-
 from surpyval import nonparametric as nonp
 from .parametric_fitter import ParametricFitter
 from .parametric import Parametric
-
-from .weibull import Weibull
 from .gumbel import Gumbel
+from .weibull import Weibull
 from .exponential import Exponential
 from .expo_weibull import ExpoWeibull
 from .normal import Normal

[…]
