### 2021-08-05 77f0647
Changed parametric fitters to flake8 standard
