+++ surpyval/parametric/fitters/mom.py
@@ -1,38 +1,40 @@
 import autograd.numpy as np
-from autograd import jacobian, hessian
-
 from scipy.optimize import minimize
+
+
+def mom_fun(params, dist, inv_trans, const, offset, moments):
+    dist_moments = dist.mom_moment_gen(*inv_trans(const(params)),
+                                       offset=offset)
+    return np.log(((np.abs(moments - dist_moments) / moments))).sum()
+
 
 def mom(model):
     """
     MOM: Method of Moments.
 
-    This is one of the simplest ways to calculate the parameters of a distribution.
-
-    This method is quick but only works with uncensored data.
+    This is one of the simplest ways to calculate the parameters of a
+    distribution. This method is quick but only works with uncensored data.
     """
     dist = model.dist
-    x, c, n, t = model.data['x'], model.data['c'], model.data['n'], model.data['t']
+    x, n = model.data['x'], model.data['n']
 
     const = model.fitting_info['const']
     inv_trans = model.fitting_info['inv_trans']
     init = model.fitting_info['init']
     offset = model.offset
-    bounds = model.bounds
 
     x_ = np.repeat(x, n)
 
     if hasattr(dist, '_mom'):
-        return {'params' : dist._mom(x_)}
+        return {'params': dist._mom(x_)}
 
     moments = np.zeros(model.k)
 
     for i in range(0, model.k):
-        moments[i] = (x_**(i+1)).mean()
+        moments[i] = (x_**(i + 1)).mean()
 
-    fun = lambda params : np.log(((np.abs(moments - dist.mom_moment_gen(*inv_trans(const(params)), offset=offset))/moments))).sum()
-    
-    res = minimize(fun, np.array(init), tol=1e-15)
+    res = minimize(mom_fun, np.array(init), tol=1e-15,
+                   args=(dist, inv_trans, const, offset, moments))
 
     params = inv_trans(const(res.x))
 
