### 2021-08-05 derrynknife/SurPyval#17 @CamDavidsonPilon
[Joss review] Test tolerance and sample sizes
I'd like to see tighter tolerances in the tests, and tests with smaller samples sizes. Currently the [TOL](https://github.com/derrynknife/SurPyval/commit/d1105f13342d0b4368a1b7197ac5d5d0788933b6) is set to 0.2 - this is quite a large margin, practically useless in some cases. The minimum sample size is 5000 - that's medium-sized data, but how well do the algorithms work for small size data (sub 100, for example).

https://github.com/openjournals/joss-reviews/issues/3484
