### 2021-08-05 derrynknife/SurPyval#18 @CamDavidsonPilon
[Joss review] Need tests for Kaplan Meier and other non-parametrics
The KM is by far the most widely used model, and most important non-parametric model. I'd like to see sufficient tests for it included.

https://github.com/openjournals/joss-reviews/issues/3484
