### 2021-08-05 derrynknife/SurPyval#19 @CamDavidsonPilon
[Joss review] Kaplan Meier underflow
For large datasets with lots of unique observation times, the `cumprod` in [KaplanMeierFitter](https://github.com/derrynknife/SurPyval/blob/master/surpyval/nonparametric/kaplan_meier.py#L9) will underflow due to the multiplication of many values less than 0.

https://github.com/openjournals/joss-reviews/issues/3484
