### 2021-08-05 derrynknife/SurPyval#20 @CamDavidsonPilon
[Joss review] Test for regression models
Similar to #18, I'd like to see tests for regression models.

https://github.com/openjournals/joss-reviews/issues/3484
