### 2021-08-05 derrynknife/SurPyval#4 @CamDavidsonPilon
Also needs docs / parametric forms on the regression models / life models.
