### 2021-08-05 derrynknife/SurPyval#21 @CamDavidsonPilon
[Joss review] tests with real-life data, and compare to other implementations
This is more relevant for testing regression models. Since real-life data never fits neatly into a "box", it is useful to compare results against other implementations (lifelines uses R's `survival` and `flexsurvreg` libraries).

openjournals/joss-reviews#3484
