### 2021-08-05 derrynknife/SurPyval#22 @CamDavidsonPilon
[Joss review] some typos in paper.md
- [x] `esimate`
- [x] `distribtuions`
- [x] `observered`
- [x] `arbitrarty`
- [x] `accomodate`

https://github.com/openjournals/joss-reviews/issues/3484
