### 2021-08-05 derrynknife/SurPyval#23 @CamDavidsonPilon
[Joss review] `scipy` should be capitalized in paper.md 
https://github.com/openjournals/joss-reviews/issues/3484
