### 2021-08-05 derrynknife/SurPyval#19 @derrynknife
I'm not convinced this is such a large risk. As the number of unique observations grow, the at risk set 'r' will also be very large which in turn makes the d/r figure small making the eventual product series between a lot of numbers that are just below 1. I ran a simulation with 1 billion unique failures:

```python

def km(r, d):
    R = 1 - (d / r)
    R[np.isnan(R)] = 0
    return R

N = 1_000_000_000
r = np.array(range(1, N+1))[::-1]
d = np.ones(N)

np.seterr(under='raise')
km(r, d)
```

This did not result in an underflow.

I can add a check for underflow and change it to a sum of the log of the (1 - d/r) value which would add some numeric stability. But I'm not sure it will ever actually be invoked.

Thoughts?
