### 2021-08-05 derrynknife/SurPyval#19 @CamDavidsonPilon
:+1: for testing. It's up to you if you want to add the check. I see no harm in using the log-transform, less some readability.
