### 2021-08-05 derrynknife/SurPyval#19 @derrynknife
Yeah, I think I was just shying away from it for readability.

Will make the change.

Thanks
