### 2021-08-05 d4df11f
Chagned tolerances for convergence testing and created small sample test of fit as well. Addresses issue 17
