+++ surpyval/tests/test_fit.py
@@ -17,14 +17,19 @@ parameter_sample_bounds = [((1, 20), (0.5, 5)),
                            ((1, 100), (1, 100)),
                            ((.1, 30), (.1, 30)),
                            ((1, 30), (.1, 10), (0.5, 1.5)),]
-FIT_SIZES = [5000, 10000, 20000, 50000, 100000, 1000000]
-TOL = 0.2
+FIT_SIZES = [1_000, 10_000, 100_000]
 
 
 def generate_mle_test_cases():
     for idx, dist in enumerate(DISTS):
         bounds = parameter_sample_bounds[idx]
         for kind in ['full', 'censored', 'truncated', 'interval']:
+            yield dist, bounds, kind
+
+def generate_small_mle_test_cases():
+    for idx, dist in enumerate(DISTS):
+        bounds = parameter_sample_bounds[idx]
+        for kind in ['full']:
             yield dist, bounds, kind
 
 
@@ -121,7 +126,8 @@ def truncate_at(x, q, where='right'):
 @pytest.mark.parametrize("dist,bounds,kind",
                          generate_mle_test_cases(),
                          ids=idfunc)
-def test_mle(dist, bounds, kind):
+def test_mle_convergence(dist, bounds, kind):
+    tol = 0.025
     for n in FIT_SIZES:
         test_params = []
         for b in bounds:
@@ -130,28 +136,55 @@ def test_mle(dist, bounds, kind):
         x = dist.random(n, *test_params)
         if kind == 'full':
             model = dist.fit(x)
-            tol = 0.1
         elif kind == 'censored':
             x, c = censor_at(x, 0.025, 'right')
-            tol = 0.1
             model = dist.fit(x, c=c)
         elif kind == 'truncated':
             x, tl, tr = truncate_at(x, 0.05, 'both')
             model = dist.fit(x, tl=tl, tr=tr)
-            tol = 0.15
         elif kind == 'interval':
             x, n = interval_censor(x)
             model = dist.fit(x=x, n=n)
-            tol = 0.15
         if model.params == []:
             continue
         fitted_params = np.array(model.params)
         max_params = np.max([fitted_params, test_params], axis=0)
         diff = np.abs(fitted_params - test_params) / max_params
-        if (diff < TOL).all():
+        # Decrease the tolerance for every parameter
+        # e.g. Weibull (2 params) tol will be 5%
+        # ExpoWeibull the tolerance will be 7.5%
+        if (diff < tol * dist.k).all():
             break
     else:
-        raise AssertionError('MLE fit not very good in %s\n' % dist.name)
+        raise AssertionError('MLE convergence not good for %s\n' % dist.name)
+
+@pytest.mark.parametrize("dist,bounds,kind",
+                         generate_small_mle_test_cases(),
+                         ids=idfunc)
+def test_mle_convergence_small(dist, bounds, kind):
+    tol = 0.03
+    for n in [100, 250, 500]:
+        test_params = []
+        for b in bounds:
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
+        x = dist.random(n, *test_params)
+        if kind == 'full':
+            model = dist.fit(x)
+        if model.params == []:
+            continue
+        fitted_params = np.array(model.params)
+        max_params = np.max([fitted_params, test_params], axis=0)
+        diff = np.abs(fitted_params - test_params) / max_params
+        # Decrease the tolerance for every parameter
+        # e.g. Weibull (2 params) tol will be 6%
+        # ExpoWeibull the tolerance will be 9%
+        if (diff < tol * dist.k).all():
+            break
+    else:
+        raise AssertionError('MLE fit for small data not good for %s\n' 
+                             % dist.name)
+
 
 
[…]
