### 2021-08-05 23b1466
Commit to close issues 22 and 23. Fixed typos in paper and capitalized SciPy
