### 2021-08-05 38e6f21
closes issue 19. Addressed risk that underflow will occur through use of cumprod. Added catch for float error and use cumsum of log transform instead
