+++ surpyval/nonparametric/kaplan_meier.py
@@ -10,8 +10,9 @@ def km(r, d):
 
     try:
         R = np.cumprod(R)
+        raise FloatingPointError
     except FloatingPointError:
-        R = np.cumsum(np.log(R))
+        R = np.cumsum(np.log(1 - (d / r)))
         R = np.exp(R)
 
     np.seterr(**old_err_state)
