### 2021-08-05 79bbfcc
Removed raise FloatingPointError that was used in testing
