### 2021-08-06 derrynknife/SurPyval#18 @derrynknife
I've created tests for the Nelson-Aalen, Kaplan-Meier, and Fleming-Harrington, and was working on the Turnbull, and my first step was to plagiarise lifelines.

I did some looking and I can't convince myself that the NPMLE implementation in lifelines is correct. Am I missing something?

If we take the midpoint for the test example in lifelines and treat it like a non-intervally censored example we get substantialy different results:

``` python
import surpyval as surv
import numpy as np

left = [1, 8, 8, 7, 7, 17, 37, 46, 46, 45]
right = [7, 8, 10, 16, 14, np.inf, 44, np.inf, np.inf, np.inf]

x = np.vstack([left, right])

c = (~np.isfinite(x[1, :])).astype(int)
x = np.where(np.isfinite(x.mean(axis=0)), x.mean(axis=0), x[0, :])

print(surpyval.KaplanMeier.fit(x=x, c=c).R)
"""
[0.9   0.8   0.7   0.6   0.5   0.5   0.375 0.375 0.375]
"""
```
The lifelines test compares this to:
```
[0.16667016, 0.33332984, 0.125, 0.375]
```
And it's inverse is:
```
[0.83332984, 0.66667016, 0.875     , 0.625     ]
```
The solution that lifelines compares to doesn't align to the Kaplan-Meier solution if the midpoints are used. The lifelines npmle seems to have the correct value of the survival function as the CDF for the last value. Am I getting something wrong?
