### 2021-08-06 derrynknife/SurPyval#18 @CamDavidsonPilon
Can you share the lifelines code used?
