### 2021-08-06 derrynknife/SurPyval#18 @CamDavidsonPilon
So `npmle` is a low-level function, and it needs to be paired with `reconstruct_survival_function`:
```
from lifelines.fitters.npmle import *
results = npmle(left, right)
sf = reconstruct_survival_function(*results)
```

This is equivalent to the (high-level API) `KaplanMeierFitter().fit_interval_censoring(left, right)`
