### 2021-08-07 fd1214a
Fixed bug where values below lowest x value in sf() for nonparametric would be -1 and therefore estimate the highest value of R.
