+++ paper.md
@@ -50,7 +50,7 @@ For the Non-Parametric estimation *SurPyval* can estimate the survival distribut
 | **Fleming-Harrington** | Non-Parametric | Yes | Right only | Left only |
 | **Turnbull** | Non-Parametric | Yes | Yes | Yes |
 
-*SurPyval* achieves this flexibility with a simple API. *SurPyval* uses a data input API, the 'xcnt' format, that can be used to define any arbitrarty combination of censored or truncated data. 'x' is the value at the observation, 'c' is the censoring flag, 'n' is the counts, and 't' is the truncation values. *SurPyval* uses the convention for the censor flag where -1 is left censored, 0 is an observed value, 1 is right censored, and 2 is intervally censored. Utilities have also been created to help users transform their data into the 'xcnt' format if they have it in another format. For example, a lot of survival data is provided in an observed and suspended format, this is where you have a list of the failure times and a list of the suspended times. E.g. Failures of [1, 2, 3, 4, 5] and suspended times of [1, 2, 3]. *SurPyval* refers to this format as the 'fs' format.
+*SurPyval* achieves this flexibility with a simple API. *SurPyval* uses a data input API, the 'xcnt' format, that can be used to define any arbitrary combination of censored or truncated data. 'x' is the value at the observation, 'c' is the censoring flag, 'n' is the counts, and 't' is the truncation values. *SurPyval* uses the convention for the censor flag where -1 is left censored, 0 is an observed value, 1 is right censored, and 2 is intervally censored. Utilities have also been created to help users transform their data into the 'xcnt' format if they have it in another format. For example, a lot of survival data is provided in an observed and suspended format, this is where you have a list of the failure times and a list of the suspended times. E.g. Failures of [1, 2, 3, 4, 5] and suspended times of [1, 2, 3]. *SurPyval* refers to this format as the 'fs' format.
 
 For Non-Parametric analysis *SurPyval* takes as input the 'xcnt' format, but the KM, NA, and FH estimators are calculated using the 'xrd' format. This format takes 'x' as the value of the observation, 'r' as the count of items at risk at each 'x', and 'd' is the number of deaths/failures at each time 'x'. Once data is in this format it is possible to compute the KM, NA, or FH estimators. For the Turnbull estimator the values of 'x', 'r', and 'd' are computed using Turnbull's algorithm; 'x', 'r', and 'd' estimates account for truncation and censoring. These values can then be used with the KM, NA, or FH estimator to get an estimate of the distribution.
 
