### 2021-08-07 derrynknife/SurPyval#18 @derrynknife
Thanks, that makes more sense, and sorry to be a pest, but I've implemented surpyval differently to avoid the following scenario.

``` python
alpha = 20
beta = 15
x = np.random.weibull(beta, 100) * alpha
n, xx = np.histogram(x, bins=20)
x = np.vstack([xx[0:-1], xx[1::]])

kmf.fit_interval_censoring(x[0, :]+1e-12, x[1, :]).plot()
ax = plt.gca()
kmf.fit_interval_censoring(x[0, :], x[1, :]).plot(color='red', ax=ax)
```

This results in catastrophic differences in the estimates. It seems that the left edge of some intervals are considered to be failing in the previous interval. This is behaviour that I wouldn't expect and therefore comparisons between lifelines and surpyval are different.

Are you sure the Turnbull intervals in lifelines are correct?
