### 2021-08-07 30e547f
Fixed bug of -1 idx in the confidence bounds as well
