### 2021-08-07 c8d1b90
Fixed handling of exactly observed events in Turnbull
