+++ surpyval/nonparametric/turnbull.py
@@ -7,6 +7,8 @@ from .fleming_harrington import fh
 
 def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
     bounds = np.unique(np.concatenate([np.unique(x), np.unique(t)]))
+    bounds = np.sort(np.concatenate([bounds, np.unique(x[c==0])]))
+    # bounds = np.sort(np.concatenate([bounds, np.repeat(np.unique(x[c==0]), 2)]))
     N = n.sum()
 
     if x.ndim == 1:
@@ -28,7 +30,7 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
     xl[c == -1] = -np.inf
     xr[c == 1] = np.inf
 
-    # bounds = bounds[0:-1]
+    # Find the count of intervals (M) and unique observation windows (N)
     M = bounds.size
     N = xl.size
 
@@ -39,7 +41,9 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
         x1, x2 = xl[i], xr[i]
         t1, t2 = tl[i], tr[i]
         if x1 == x2:
-            alpha[i, :] = (bounds == x1).astype(int) * n[i]
+            # Find first index of repeated value
+            idx = np.where(bounds == x1)[0][0]
+            alpha[i, idx] = n[i]
         elif x2 == np.inf:
             alpha[i, :] = ((bounds > x1) & (bounds <= x2)).astype(int) * n[i]
         else:
@@ -64,11 +68,14 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
 
     old_err_state = np.seterr(all='ignore')
 
-    while (iters < 1000) & (not np.allclose(p, p_prev)):
+    while (iters < 1000) & (not np.allclose(p,
+                                            p_prev,
+                                            rtol=1e-30,
+                                            atol=1e-30)):
         p_prev = p
         iters += 1
         ap = alpha * p
-        # Observed deaths
+        # Observed deaths (left, right, or intervally 'observed')
         mu = alpha * ap / ap.sum(axis=1, keepdims=True)
         # Expected additional failures due to truncation
         nu = (n * (1 - beta) * (1 - beta) * p
@@ -78,27 +85,33 @@ def turnbull(x, c, n, t, estimator='Fleming-Harrington'):
         d = (nu + mu).sum(axis=0)
         # M total observed and unobserved failures.
         M = (nu + mu).sum()
-        # Risk set, i.e the number of items at risk at each value x
+        # Risk set, i.e the number of items at risk at immediately before x
         r = M - d.cumsum() + d
-        # Find the reliability using the deaths and risk set.
+        # Find the survival function values (R) using the deaths and risk set
         R = func(r, d)
         # Calculate the probability mass in each interval
         p = np.abs(np.diff(np.hstack([[1], R])))
 
-        # The 'official' way to do it which is equivalent to the Kaplan-Meier
+        # The 'official' way to do it which is equivalent to using KM
         # p = (nu + mu).sum(axis=0)/(nu + mu).sum()
 
     # Remove the -Inf and the Inf values.
     x = bounds[1:-1]
     r = r[1:-1]
     d = d[1:-1]
-    R = R[1:-1]
+    R = R[0:-2]
+    R_upper = R[0:-2]
+    R_lower = R[1:-1]
 
     out = {}
     out['x'] = x
     out['r'] = r
     out['d'] = d
     out['R'] = R
+    out['R_upper'] = R_upper
+    out['R_lower'] = R_lower
+    out['alpha'] = alpha
+    out['bounds'] = bounds
 
[…]
