### 2021-08-07 7869c2c
Fixed bug that made zi or lfp models fail. Was created in changes made to meet flake8 standards
