### 2021-08-07 2c4b759
beta.py converted to flake8 standards
