+++ surpyval/parametric/beta.py
@@ -1,26 +1,11 @@
-import types
 import autograd.numpy as np
-from autograd import jacobian
 from scipy.stats import uniform
-from scipy.optimize import approx_fprime
-from scipy.special import betainc, betaincinv
+from scipy.special import betaincinv
 from scipy.special import gamma as gamma_func
-from scipy.special import gammainc, gammaincinv
-from autograd_gamma import gammainc as agammainc
 from autograd.scipy.special import beta as abeta
 from autograd_gamma import betainc as abetainc
-from scipy.special import ndtri as z
-from scipy.linalg import lstsq
-
-import warnings
-
-from scipy.optimize import minimize
-from scipy.stats import pearsonr
-
-import surpyval
-from surpyval import parametric as para
-from surpyval import nonparametric as nonp
 from surpyval.parametric.parametric_fitter import ParametricFitter
+
 
 class Beta_(ParametricFitter):
     def __init__(self, name):
@@ -29,14 +14,14 @@ class Beta_(ParametricFitter):
         self.bounds = ((0, None), (0, None),)
         self.support = (0, 1)
         self.plot_x_scale = 'linear'
-        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002,
+                        0.003, 0.005, 0.01, 0.02, 0.03, 0.05,
+                        0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8,
+                        0.9, 0.95, 0.99, 0.999, 0.9999]
         self.param_names = ['alpha', 'beta']
         self.param_map = {
-            'alpha' : 0,
-            'beta'  : 1
+            'alpha': 0,
+            'beta': 1
         }
 
     def _parameter_initialiser(self, x, c=None, n=None):
@@ -53,13 +38,14 @@ class Beta_(ParametricFitter):
         Survival (or reliability) function for the Beta Distribution:
 
         .. math::
-            R(x) = 1 - \int_{0}^{x}t^{\alpha-1}\left (1 - t \right )^{\beta -1}dt
+            R(x) = 1 - \int_{0}^{x}t^{\alpha-1}\left (1 - t \right )^
+            {\beta - 1}dt
 
         Parameters
         ----------
 
         x : numpy array or scalar
-            The values at which the function will be calculated 
+            The values at which the function will be calculated
         alpha : numpy array or scalar
             One shape parameter for the Beta distribution
         beta : numpy array or scalar
@@ -68,7 +54,7 @@ class Beta_(ParametricFitter):
         Returns
         -------
 
-        df : scalar or numpy array 
+        df : scalar or numpy array
             The value(s) of the reliability function at x.
 
         Examples
@@ -84,7 +70,8 @@ class Beta_(ParametricFitter):
     def cs(self, x, X, alpha, beta):
         r"""
 
-        Conditional survival (or reliability) function for the Beta Distribution:
+        Conditional survival (or reliability) function for the Beta
+        Distribution:
 
         .. math::
             R(x, X) = \frac{R(x + X)}{R(X)}
@@ -93,7 +80,7 @@ class Beta_(ParametricFitter):
         ----------
 
         x : numpy array or scalar
-            The values at which the function will be calculated 
+            The values at which the function will be calculated
         X : numpy array or scalar
             The value(s) at which each value(s) in x was known to have survived
[…]
