+++ surpyval/parametric/parametric.py
@@ -1,17 +1,14 @@
 import re
-from autograd import jacobian
-from autograd import grad
 import autograd.numpy as np
+import matplotlib.pyplot as plt
+from autograd import jacobian
 from scipy.stats import uniform
-
 from surpyval import round_sig, fs_to_xcn
 from scipy.special import ndtri as z
 from surpyval import nonparametric as nonp
 from copy import deepcopy, copy
-
-import matplotlib.pyplot as plt
-from scipy.optimize import approx_fprime
 from matplotlib.ticker import FixedLocator
+
 
 def _round_vals(x):
     not_different = True
@@ -22,15 +19,15 @@ def _round_vals(x):
         i += 1
     return x_ticks
 
+
 class Parametric():
     """
-    Result of ``.fit()`` or ``.from_params()`` method for every parametric surpyval 
-    distribution.
-    
-    Instances of this class are very useful when a user needs the other functions 
-    of a distribution for plotting, optimizations, monte carlo analysis and numeric 
-    integration.
+    Result of ``.fit()`` or ``.from_params()`` method for every parametric
+    surpyval distribution.
 
+    Instances of this class are very useful when a user needs the other
+    functions of a distribution for plotting, optimizations, monte carlo
+    analysis and numeric integration.
     """
     def __init__(self, dist, method, data, offset, lfp, zi):
         self.dist = dist
@@ -50,22 +47,22 @@ class Parametric():
             else:
                 bounds = ((None, None), *bounds)
 
-            param_map = {k : v + 1 for k, v in param_map.items()}
-            param_map.update({'gamma' : 0})
-            self.k +=1
+            param_map = {k: v + 1 for k, v in param_map.items()}
+            param_map.update({'gamma': 0})
+            self.k += 1
         else:
             self.gamma = 0
 
         if lfp:
             bounds = (*bounds, (0, 1))
-            param_map.update({'p' : len(param_map) + 1})
+            param_map.update({'p': len(param_map) + 1})
             self.k += 1
         else:
             self.p = 1
 
         if zi:
             bounds = (*bounds, (0, 1))
-            param_map.update({'f0' : len(param_map) + 1})
+            param_map.update({'f0': len(param_map) + 1})
             self.k += 1
         else:
             self.f0 = 0
@@ -73,18 +70,20 @@ class Parametric():
         self.bounds = bounds
         self.param_map = param_map
 
-
     def __repr__(self):
-        param_string = '\n'.join(['{:>10}'.format(name) + ": " + str(p) for p, name in zip(self.params, self.dist.param_names)])
+        param_string = '\n'.join(['{:>10}'.format(name)
+                                  + ": " + str(p) for p, name in
+                                  zip(self.params, self.dist.param_names)])
+
         if hasattr(self, 'params'):
             out = ('Parametric SurPyval Model'
                    + '\n========================='
                    + '\nDistribution        : {dist}'
                    + '\nFitted by           : {method}'
-                   ).format(dist=self.dist.name, 
+                   ).format(dist=self.dist.name,
                             method=self.method)
             if self.offset:
-                out += '\nOffset (gamma)      : {gamma}'.format(gamma=self.gamma)
+                out += '\nOffset (gamma)      : {g}'.format(g=self.gamma)
 
             if self.lfp:
                 out += '\nMax Proportion (p)  : {p}'.format(p=self.p)
@@ -102,20 +101,24 @@ class Parametric():
     def sf(self, x):
         r"""
 
-        Surival (or Reliability) function for a distribution using the parameters found in the ``.params`` attribute.
+        Surival (or Reliability) function for a distribution using the
+        parameters found in the ``.params`` attribute.
 
         Parameters
         ----------
 
         x : array like or scalar
[…]
