### 2021-08-07 28d38b3
mixture_model.py converted to flake8 standard
