+++ surpyval/parametric/mixture_model.py
@@ -1,195 +1,201 @@
+import surpyval
 import autograd.numpy as np
 from scipy.linalg import inv
-from scipy.stats import uniform
 from scipy.optimize import minimize
 from scipy.special import ndtri as z
 
 from autograd import jacobian, hessian
 from surpyval import parametric as para
 
+
 class MixtureModel():
-	"""
-	Generalised from algorithm found here
-	https://www.sciencedirect.com/science/article/pii/S0307904X12002545
-	"""
-	def __init__(self, x, dist=para.Weibull, **kwargs):
-		self.m = kwargs.pop('m', 2)
-		raw_data = {}
-		c = kwargs.pop('c', None)
-		n = kwargs.pop('n', None)
-		raw_data['x'] = x
-		raw_data['c'] = c
-		raw_data['n'] = n
-		self.raw_data = raw_data
-		self.dist = dist
+    """
+    Generalised from algorithm found here
+    https://www.sciencedirect.com/science/article/pii/S0307904X12002545
+    """
+    def __init__(self, x, dist=para.Weibull, **kwargs):
+        self.m = kwargs.pop('m', 2)
+        raw_data = {}
+        c = kwargs.pop('c', None)
+        n = kwargs.pop('n', None)
+        raw_data['x'] = x
+        raw_data['c'] = c
+        raw_data['n'] = n
+        self.raw_data = raw_data
+        self.dist = dist
 
-		x, c, n = surpyval.xcn_handler(x, c, n)
-		assert len(x) > self.m * (self.dist.k + 1)
+        x, c, n = surpyval.xcn_handler(x, c, n)
+        assert len(x) > self.m * (self.dist.k + 1)
 
-		self.x = x
-		self.c = c
-		self.n = n
+        self.x = x
+        self.c = c
+        self.n = n
 
-		self.N = n.sum()
-		splits_x = np.array_split(x, self.m)
-		splits_c = np.array_split(c, self.m)
-		splits_n = np.array_split(n, self.m)
-		params = np.zeros(shape=(self.m, self.dist.k))
+        self.N = n.sum()
+        splits_x = np.array_split(x, self.m)
+        splits_c = np.array_split(c, self.m)
+        splits_n = np.array_split(n, self.m)
+        params = np.zeros(shape=(self.m, self.dist.k))
 
-		for i in range(self.m):
-			params[i, :] = self.dist.fit(x=splits_x[i], c=splits_c[i], 
-										 n=splits_n[i]).params
-		self.params = params
-		self.w = np.ones(shape=(self.m)) / self.m
-		self.p = np.ones(shape=(self.m, len(self.x))) / self.m
+        for i in range(self.m):
+            params[i, :] = self.dist.fit(x=splits_x[i], c=splits_c[i],
+                                         n=splits_n[i]).params
+        self.params = params
+        self.w = np.ones(shape=(self.m)) / self.m
+        self.p = np.ones(shape=(self.m, len(self.x))) / self.m
 
-	def Q(self, params):
-		params = params.reshape(self.m, self.dist.k)
-		f = np.zeros_like(self.p)
-		for i in range(self.m):
-			like = self.dist.like(self.x, self.c, self.n, *params[i])
-			like += surpyval.TINIEST
-			like = np.where(like < 1, like, 1)
-			like = np.log(like)
-			like = np.multiply(self.n, like)
+    def Q(self, params):
+        params = params.reshape(self.m, self.dist.k)
+        f = np.zeros_like(self.p)
+        for i in range(self.m):
+            like = self.dist.like(self.x, self.c, self.n, *params[i])
[…]
