### 2021-08-07 a160e99
separated nonparametric tests from parametric fit tests
