+++ surpyval/tests/test_np.py
@@ -1,0 +1,181 @@
+import pytest
+import numpy as np
+import lifelines
+import surpyval
+
+def right_censor(x, tl, frac):
+    c = np.random.binomial(1, frac, x.shape)
+    x_out = np.copy(x)
+    for i, (trunc, value) in enumerate(zip(tl, x)):
+        if c[i] == 0:
+            continue
+        if np.isfinite(trunc):
+            x_out[i] = np.random.uniform(trunc, value, 1)
+        else:
+            x_out[i] = value - np.abs(value * np.random.uniform(0, 1, 1))
+    return x_out, c
+
+def left_truncate(x, dist, frac, params):
+    t = np.random.binomial(1, frac, x.shape)
+    # Find a lower value
+    tl = np.where((t == 1) & (x > 0), x * np.random.uniform(0, 1, x.size), -np.inf)
+    tl = np.where((t == 1) & (x < 0), x - np.abs(x * np.random.uniform(0, 1, x.size)), tl)
+    drop_due_to_truncation = dist.ff(tl, *params)
+    drop_due_to_truncation[~np.isfinite(drop_due_to_truncation)] = 0
+    
+    keep = np.ones_like(x)
+    for i, p in enumerate(drop_due_to_truncation):
+        if p == 0:
+            continue
+        else:
+            keep[i] = np.random.binomial(1, p)
+
+    mask = keep == 1
+    tl = tl[mask]
+    x = x[mask]
+    return x, tl
+
+def test_kaplan_meier_against_lifelines():
+    kmf = lifelines.KaplanMeierFitter()
+    for i in range(100):
+        test_params = []
+        for b in ((1, 100), (0.5, 20)):
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
+        x = surpyval.Weibull.random(int(np.random.uniform(2, 1000, 1)), *test_params)
+        n = np.ones_like(x) * int(np.random.uniform(1, 5))
+        ll_est = kmf.fit(x, weights=n).predict(x).values
+        surp_est = surpyval.KaplanMeier.fit(x, n=n).sf(x)
+        if not np.allclose(ll_est, surp_est, 1e-15):
+            raise AssertionError('Kaplan-Meier different to lifelines?!')
+
+def test_kaplan_meier_censored_against_lifelines():
+    kmf = lifelines.KaplanMeierFitter()
+    for i in range(100):
+        test_params = []
+        for b in ((1, 100), (0.5, 20)):
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
+        x = surpyval.Weibull.random(int(np.random.uniform(2, 1000, 1)), *test_params)
+        c = np.random.binomial(1, np.random.uniform(0, 1, 1), x.shape)
+        x = x - np.abs(x * np.random.uniform(0, 1, x.shape))
+        n = np.ones_like(x) * int(np.random.uniform(1, 5))
+        ll_est = kmf.fit(x, 1-c, weights=n).predict(x).values
+        surp_est = surpyval.KaplanMeier.fit(x, c=c, n=n).sf(x)
+        if not np.allclose(ll_est, surp_est, 1e-15):
+            raise AssertionError('Kaplan-Meier different to lifelines?!')
+
+
+def test_kaplan_meier_censored_and_truncated_against_lifelines():
+    kmf = lifelines.KaplanMeierFitter()
+    for i in range(100):
+        test_params = []
+        for b in ((1, 100), (0.5, 20)):
+            test_params.append(np.random.uniform(*b))
+        test_params = np.array(test_params)
[…]
