### 2021-08-07 3520965
gamma.py changed to flake8 standard
