+++ surpyval/parametric/gamma.py
@@ -1,21 +1,16 @@
 import autograd.numpy as np
 from autograd import jacobian
 from scipy.stats import uniform
-from scipy.optimize import approx_fprime
-from scipy.special import gamma as gamma_func
 from autograd.scipy.special import gamma as agamma
-from scipy.special import gammainc, gammaincinv
+from scipy.special import gammaincinv
 from autograd_gamma import gammainc as agammainc
 from scipy.special import ndtri as z
-
 from scipy.optimize import minimize
 from scipy.stats import pearsonr
-
 import warnings
-
-from surpyval import parametric as para
 from surpyval import nonparametric as nonp
 from surpyval.parametric.parametric_fitter import ParametricFitter
+
 
 class Gamma_(ParametricFitter):
     r"""
@@ -33,14 +28,14 @@ class Gamma_(ParametricFitter):
         self.bounds = ((0, None), (0, None),)
         self.support = (0, np.inf)
         self.plot_x_scale = 'linear'
-        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002, 
-            0.003, 0.005, 0.01, 0.02, 0.03, 0.05, 
-            0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 
-            0.9, 0.95, 0.99, 0.999, 0.9999]
+        self.y_ticks = [0.0001, 0.0002, 0.0003, 0.001, 0.002,
+                        0.003, 0.005, 0.01, 0.02, 0.03, 0.05,
+                        0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8,
+                        0.9, 0.95, 0.99, 0.999, 0.9999]
         self.param_names = ['alpha', 'beta']
         self.param_map = {
-            'alpha' : 0,
-            'beta'  : 1
+            'alpha': 0,
+            'beta': 1
         }
 
     def _parameter_initialiser(self, x, c=None, n=None, offset=False):
@@ -62,13 +57,14 @@ class Gamma_(ParametricFitter):
         Surival (or Reliability) function for the Gamma Distribution:
 
         .. math::
-            R(x) = 1 - \frac{\gamma \left ( \alpha, \beta x \right )}{\Gamma \left ( \alpha \right )}
+            R(x) = 1 - \frac{\gamma \left ( \alpha, \beta x \right )
+            }{\Gamma \left ( \alpha \right )}
 
         Parameters
         ----------
 
         x : numpy array or scalar
-            The values at which the function will be calculated 
+            The values at which the function will be calculated
         alpha : numpy array or scalar
             The shape parameter for the Gamma distribution
         beta : numpy array or scalar
@@ -78,7 +74,7 @@ class Gamma_(ParametricFitter):
         Returns
         -------
 
-        sf : scalar or numpy array 
+        sf : scalar or numpy array
             The value(s) for the survival function at each x
 
         Examples
@@ -114,7 +110,7 @@ class Gamma_(ParametricFitter):
         Returns
         -------
 
-        cs : scalar or numpy array 
+        cs : scalar or numpy array
             the conditional survival probability.
 
         Examples
@@ -134,13 +130,14 @@ class Gamma_(ParametricFitter):
         CDF (or unreliability or failure) function for the Gamma Distribution:
 
         .. math::
-            F(x) = \frac{\gamma \left ( \alpha, \beta x \right )}{\Gamma \left ( \alpha \right )}
+            F(x) = \frac{\gamma \left ( \alpha, \beta x \right )}
+            {\Gamma \left ( \alpha \right )}
 
         Parameters
         ----------
 
         x : numpy array or scalar
-            The values at which the function will be calculated 
[…]
