### 2021-08-07 f3a264a
Made nonparametric tests at random values around the numbers input. Makes it test above and below range of random numbers
