### 2021-08-07 derrynknife/SurPyval#17 @derrynknife
Fair, I changed it to 20% because the git workflows seemed to not converge as well as my local machine. I've changed it back to much finer tolerances and split the convergence into two tests. A large scale convergence and a small sample convergence.

commit d4df11ff7c13aa372ee843d6585751282706de98
