### 2021-08-07 derrynknife/SurPyval#19 @derrynknife
Created catch of floating point errors and use log in the event of such an error.

See:
653fbb1544dcd9b4dc7c40fcb707531f70ec4465
