### 2021-08-07 derrynknife/SurPyval#23 @derrynknife
Capitalized in paper.

23b14665ea6873e6e19e99cd4218301dd2fc8b8c
