### 2021-08-07 derrynknife/SurPyval#21 @derrynknife
I only completed the regression model fitters after writing the paper.  And in the paper I make no claims about regression models. If I remove all references to regression models, and submit a major change in future, can we exclude the regression features from this review?
