### 2021-08-07 derrynknife/SurPyval#20 @derrynknife
See commend in [issue 12](https://github.com/derrynknife/SurPyval/issues/21). Same question applies here.
