### 2021-08-07 derrynknife/SurPyval#18 @CamDavidsonPilon
> Are you sure the Turnbull intervals in lifelines are correct?

No lol - I only know they are correct up-to the unit tests in lifelines.

The data you presented is an interesting case I'll have to explore more. I would suggest comparing results to `icenreg` in R if you can. (Seeing your written tests now - I suggest comparing vs R's implementations as well as lifelines)
