### 2021-08-07 derrynknife/SurPyval#21 @CamDavidsonPilon
I'm fine with dropping #20, but then I would still suggest testing real-life data in non-regression tests and comparing vs R and/or lifelines. This can also test the robustness of your `_parameter_initialiser` methods (since we'll be testing it against different-distributed data points)
