### 2021-08-08 4086c12
Added tests for nonparametric. Compares to R, where possible.
