+++ surpyval/tests/test_np.py
@@ -185,5 +185,170 @@ def test_fleming_harrington_case():
     assert np.allclose(H, H_est, 1e-15)
 
 
+def test_turnbull_with_R_1():
+    """
+    R's icenReg solution, from code:
+
+    ``` r
+    left <-  c(1, 8, 8, 7, 7, 17, 37, 46, 46, 45)
+    right <- c(7, 8, 15, 16, 14, Inf, 44, Inf, Inf, Inf)
+
+    df <- data.frame(cbind(left, right))
+
+    r_model <- ic_np(cbind(left, right) ~ 0, df)
+
+    getSCurves(r_model)
+    ```
+    """
+    x_test = [1, 7, 8, 14, 44, np.inf]
+    sf_test = [1., 0.9, 0.7, 0.5, 0.375, 0]
+
+    left = np.array([1, 8, 8, 7, 7, 17, 37, 46, 46, 45])
+    right = np.array([7, 8, 10, 16, 14, np.inf, 44, np.inf, np.inf, np.inf])
+
+    model = surpyval.Turnbull.fit(xl=left,
+                              xr=right,
+                              turnbull_estimator='Kaplan-Meier')
+
+    assert np.allclose(model.sf(x_test), sf_test, 1e-5)
 
 
+def test_turnbull_with_R_2():
+    """
+    R's icenReg solution, from code:
+
+    ``` r
+    left <- c(1, 76, 288, 501, 579, 667, 829, 920, 1071)
+    right <- c(71, 169, 344, 504, 579, 754, 829, 971, Inf)
+
+    df <- data.frame(cbind(left, right))
+
+    r_model <- ic_np(cbind(left, right) ~ 0, df)
+
+    getSCurves(r_model)
+    ```
+    """
+    x_test = [1, 71, 169, 344, 504, 579, 754, 829, 971, np.inf]
+    sf_test = [1.0, 0.8888889, 0.7777778, 0.6666667, 0.5555556, 0.4444444,
+              0.3333333, 0.2222222, 0.1111111, 0.0]
+
+
+    left = [1, 76, 288, 501, 579, 667, 829, 920, 1071]
+    right = [71, 169, 344, 504, 579, 754, 829, 971, np.inf]
+
+    model = surpyval.Turnbull.fit(xl=left,
+                                  xr=right,
+                                  turnbull_estimator='Kaplan-Meier')
+
+    assert np.allclose(model.sf(x_test), sf_test, 1e-5)
+
+[0.875, 0.75, 0.625, 0.5, 0.375, 0.25, 0.125, 0.0]
+
+def test_turnbull_with_R_3():
+    """
+    R's icenReg solution, from code:
+
+    ``` r
+    left <- seq(1, 10)
+    right <- seq(2, 11)
+
+    df <- data.frame(cbind(left, right))
+
+    t <- ic_np(cbind(left, right) ~ 0, df)
+
+    getSCurves(t)
+    ```
+    """
+    x_test = np.array(range(1, 12))
+    sf_test = [1., 0.9, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1, 0.]
+    left = np.array(range(1, 11))
+    right = left + 1
+
+
+    model = surpyval.Turnbull.fit(xl=left,
+                                  xr=right,
+                                  turnbull_estimator='Kaplan-Meier')
+
+    assert np.allclose(model.sf(x_test), sf_test, 1e-5)
+
+
+
+
+def test_kaplan_meier_with_R():
+    """
+    R's survreg() solutions.
+    """
+    # Test 1
[…]
