### 2021-08-08 derrynknife/SurPyval#18 @derrynknife
> No lol - I only know they are correct up-to the unit tests in lifelines.

No probs. SurPyval doesn't actually find the intervals, it just creates an interval for between every observed value and every truncated value and brute forces the solution. This implementation will result in a few additional calculations but it makes the code more readable and was easier to develop.
