### 2021-08-08 457c77e
Docs images updates, and removed reference to regression for JOSS review and release
