### 2021-08-09 0e16b1c
changed default plotting point to Fleming-Harrington in _parameter_initializer and it improved convergence performance. amazing.
