### 2021-08-09 d7e17b1
changed _parameter_initializer to check for infs and changed to average. Better performance.
