+++ surpyval/parametric/exponential.py
@@ -34,9 +34,9 @@ class Exponential_(ParametricFitter):
         }
 
     def _parameter_initialiser(self, x, c=None, n=None, offset=False):
-        x, c, n = xcn_handler(x, c, n)
-        c = (c == 0).astype(np.int64)
-        rate = (n * c).sum()/x.sum()
+        # x, c, n = xcn_handler(x, c, n)
+        # c = (c == 0).astype(np.int64)
+        rate = 1./x[np.isfinite(x)].mean()
         if offset:
             return np.min(x) - (np.max(x) - np.min(x))/10., rate
         else:
