### 2021-08-09 e3da8d7
added tests against real data (from lifelines datasets)
