+++ surpyval/tests/test_real_data.py
@@ -1,0 +1,344 @@
+import pytest
+import numpy as np
+import lifelines
+import surpyval as surv
+
+from collections import namedtuple
+from lifelines.datasets import *
+from reliability.Fitters import *
+
+# Datasets in x, c, n: as namedtuples
+SurvivalData = namedtuple('SurvivalData', ['x', 'c', 'n', 'name'])
+IntervalSurvivalData = namedtuple('IntervalSurvivalData',
+                                  ['left', 'right', 'name'])
+
+# Canadian Senators
+df = load_canadian_senators()
+x = df['diff_days'].values
+c = 1 - df['observed'].astype(int)
+zero_idx = (x == 0)
+x = x[~zero_idx]
+c = c[~zero_idx]
+n = np.ones_like(x)
+canadian_senators = SurvivalData(x, c, n, 'canadian_senators')
+
+# Bacteria...
+df = load_c_botulinum_lag_phase()
+left = df['lower_bound_days']
+right = df['upper_bound_days']
+bacteria = IntervalSurvivalData(left, right, 'bacteria')
+
+# Political Durations
+df = load_dd()
+x = df['duration'].values
+c = 1 - df['observed'].astype(int)
+n = np.ones_like(x)
+politics = SurvivalData(x, c, n, 'politics')
+
+df = load_diabetes()
+left = df['left']
+right = df['right']
+diabetes = IntervalSurvivalData(left, right, 'diabetes')
+
+# ???
+df = load_g3()
+x = df['time'].values
+c = 1 - df['event'].astype(int)
+n = np.ones_like(x)
+g3 = SurvivalData(x, c, n, 'g3')
+
+# ???
+df = load_gbsg2()
+x = df['time'].values
+c = 1 - df['cens'].astype(int)
+n = np.ones_like(x)
+gbsg2 = SurvivalData(x, c, n, 'gbsg2')
+
+# holl_molly_polly...??!!!?!?!?!?!
+df = load_holly_molly_polly()
+x = df['T'].values
+c = np.zeros_like(x)
+n = np.ones_like(x)
+holly_molly_polly = SurvivalData(x, c, n, 'holly_molly_polly')
+
+# kidney
+df = load_kidney_transplant()
+x = df['time'].values
+c = 1 - df['death'].astype(int)
+n = np.ones_like(x)
+kidney = SurvivalData(x, c, n, 'kidney')
+
+# larynx
+df = load_larynx()
+x = df['time'].values
+c = np.zeros_like(x)
+n = np.ones_like(x)
+larynx = SurvivalData(x, c, n, 'larynx')
+
+# leukemia
+df = load_leukemia()
+x = df['t'].values
+c = 1 - df['status'].astype(int)
+n = np.ones_like(x)
+leukemia = SurvivalData(x, c, n, 'leukemia')
+
+# lung
+df = load_lung()
+x = df['time'].dropna()
+c = np.zeros_like(x)
+n = np.ones_like(x)
+lung = SurvivalData(x, c, n, 'lung')
+
+# lupus
+df = load_lupus()
+x = df['time_elapsed_between_estimated_onset_and_diagnosis_(months)'].dropna()
+c = 1 - df['dead'].astype(int)
+n = np.ones_like(x)
+lupus = SurvivalData(x, c, n, 'lupus')
+
+# lymph
+df = load_lymph_node()
+x = df['survtime'].dropna()
+c = 1 - df['censdead'].astype(int)
+n = np.ones_like(x)
+lymph = SurvivalData(x, c, n, 'lymph')
+
+# lymphoma
+df = load_lymphoma()
+x = df['Time'].dropna()
+c = 1 - df['Censor'].astype(int)
[…]
