### 2021-08-09 3bf2502
changed test_fit.py to flake8 standard
