+++ surpyval/tests/test_fit.py
@@ -1,71 +1,68 @@
 import pytest
 import numpy as np
-import lifelines
-import surpyval
 from surpyval import (
     Gumbel, Normal, Weibull, LogNormal,
-    Logistic, LogLogistic, Uniform, 
+    Logistic, LogLogistic, Uniform,
     Gamma, Beta, ExpoWeibull)
 
-DISTS = [Gumbel, Normal, Weibull, LogNormal, 
-         Logistic, LogLogistic, Uniform, 
+DISTS = [Gumbel, Normal, Weibull, LogNormal,
+         Logistic, LogLogistic, Uniform,
          Beta, ExpoWeibull, Gamma]
-parameter_sample_bounds = [((1, 20), (0.5, 5)),
-                           ((1, 100), (0.5, 100)),
-                           ((1, 100), (0.5, 20)),
-                           ((1, 3), (0.2, 1)),
-                           ((1, 100), (0.5, 20)),
-                           ((1, 100), (0.5, 20)),
-                           ((1, 100), (1, 100)),
-                           ((.1, 30), (.1, 30)),
-                           ((1, 30), (.1, 10), (0.5, 1.5)),
-                           ((1, 30), (.1, 10)),]
-FIT_SIZES = [1_000, 10_000]#, 100_000]
+parameter_sample_random_bounds = [((1, 20), (0.5, 5)),
+                                  ((1, 100), (0.5, 100)),
+                                  ((1, 100), (0.5, 20)),
+                                  ((1, 3), (0.2, 1)),
+                                  ((1, 100), (0.5, 20)),
+                                  ((1, 100), (0.5, 20)),
+                                  ((1, 100), (1, 100)),
+                                  ((.1, 30), (.1, 30)),
+                                  ((1, 30), (.1, 10), (0.5, 1.5)),
+                                  ((1, 30), (.1, 10))]
+FIT_SIZES = [1_000, 10_000, 100_000]
 
 
 def generate_mle_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
+        random_bounds = parameter_sample_random_bounds[idx]
         for kind in ['full', 'censored', 'truncated', 'interval']:
-            yield dist, bounds, kind
+            yield dist, random_bounds, kind
+
 
 def generate_small_mle_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
+        random_bounds = parameter_sample_random_bounds[idx]
         for kind in ['full']:
-            yield dist, bounds, kind
+            yield dist, random_bounds, kind
 
 
 def generate_mpp_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
+        random_bounds = parameter_sample_random_bounds[idx]
         for rr in ['x', 'y']:
-            yield dist, bounds, rr
-
+            yield dist, random_bounds, rr
 
 
 def generate_mom_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
-        yield dist, bounds
-
+        random_bounds = parameter_sample_random_bounds[idx]
+        yield dist, random_bounds
 
 
 def generate_mps_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
-        yield dist, bounds
+        random_bounds = parameter_sample_random_bounds[idx]
+        yield dist, random_bounds
 
 
 def generate_mse_test_cases():
     for idx, dist in enumerate(DISTS):
-        bounds = parameter_sample_bounds[idx]
-        yield dist, bounds
[…]
