### 2021-08-09 derrynknife/SurPyval#21 @derrynknife
Added tests to lifelines data sets. See e3da8d782ea5e8f79d5c120122de84af89339586
