### 2021-08-09 #3484 @derrynknife
@CamDavidsonPilon. Thanks for the thorough review, I've made requested changes as per the raised issues. Although happy to add more if you think the approach is incorrect.

But the tests directory now include tests against the datasets of lifelines and compares results. I also do a comparison to @MatthewReid854's `reliability` package with other distributions not implemented in `lifelines`. `surpyval` has the same performance within a tolerance as `reliability` (and in some cases exceeds as I think `reliability` needs to be checked see [here](https://github.com/MatthewReid854/reliability/issues/21) ).
