### 2021-08-09 c2d88f5
fixed bug in __repr__ so that even if a fit fails it will still print
