### 2021-08-09 #3484 @derrynknife
Also, worth saying that some more real life data application examples can be seen [here](https://surpyval.readthedocs.io/en/latest/applications.html)
