### 2021-08-09 48f00f7
created xcn_to_fsl function
