### 2021-08-09 b2288b6
Changed interval censoring plot since bug fixes now show it is a good fit
