### 2021-08-09 014fba8
Added comment about truncation and censoring tp Gompertz example. This should explain the divergence of the confidence bounds in the probability plot.
