### 2021-08-09 dfabf81
Changed behaviour of mle fitter to use initial values if mle fails
