### 2021-08-09 79e8ae1
Fixed bug where initial guess of offset was poor due to handling of inf or out of support x values
