### 2021-08-09 #3484 @dfm
@derrynknife: I've opened a pull request with some edits to the paper. Can you take a look at that and once you've merged, please take the following steps:

1. Comment `@whedon generate pdf` on this thread and read through the manuscript to make sure that you're happy with it (it's hard to make changes later!), especially your name and affiliation.
3. Create an archived release of that version of the software (using Zenodo or something similar). Please make sure that the metadata (title and author list) exactly match the paper. Then report the DOI of the release back to this thread.

Let me know if you have questions or run into any issues!
