### 2021-08-09 #3484 @derrynknife
@dfm I've incremented the version number to 0.10.0 and uploaded to zenodo.

https://doi.org/10.5281/zenodo.5173006
