### 2021-08-10 #3484 @dfm
@derrynknife: looks good, but could you clean up the Zenodo archive a little? It has all sorts of temporary files and the git archive which we don't want. One option would be to download a zip file of a snapshot from GitHub itself and upload that directly to Zenodo. Thanks!
