### 2021-08-10 #3484 @derrynknife
@dfm. I didn't even think of that **facepalm**

I've re-uploaded it as per your recommendation, the new DOI is:

https://doi.org/10.5281/zenodo.5177222
